### 2021-09-21 #3744 @whedon
[PRE REVIEW]: specieshindex: How scientifically popular is a species?
**Submitting author:** @jessicatytam (<a href="http://orcid.org/0000-0003-3655-1974">Jessica Tam</a>)
**Repository:** <a href="https://github.com/jessicatytam/specieshindex" target ="_blank">https://github.com/jessicatytam/specieshindex</a>
**Version:** 0.3.1
