### 2021-09-23 #3744 @Bisaloo
@jessicatytam, while I'm looking for reviewers, could you please have a look at the missing DOI detected by whedon in [this comment](https://github.com/openjournals/joss-reviews/issues/3744#issuecomment-924010397)?
