### 2021-09-23 #3744 @Bisaloo
@ahasverus, would you be interested in reviewing this paper & R package for JOSS?
