### 2021-09-23 #3744 @Bisaloo
:wave: @dombennett, you signaled interest in reviewing papers for JOSS? Would you have time to review this paper & package that matches the research interests you listed.
