+++ paper.md
@@ -41,7 +41,7 @@ Over the last decade, the quantification of research biases have gained populari
 
 The package `specieshindex` connects to the Scopus and Web of Science databases (interdisciplinary broad-range databases of academic literature) and extracts citation records of relevant publications, as identified via a search query. It returns blibliometric information including the publication title, number of citations, publication type, etc. The binomial or genus names of the species should be used in the search query, instead of their common names, since the latter are less specific and may refer to more than 1, and even unrelated species. An example of such is 'pig', which can refer to the domestic pig (*Sus scrofa* - an ungulate), or the guinea pig (*Cavia porcellus* - a rodent). Although this package also connects to Bielefeld Academic Search Engine (BASE), article extraction is not available, hence *h*-index calculation is unavailable with BASE.
 
-There are two types of functions that connect to the literature databases. The count functions, e.g. `CountSpT()` and `CountSpTAK()` for species, and `CountGenusT()` and `CountGenusTAK()` for genus, return the total publication count without extracting any citation records. Whereas the fetch functions, e.g. `FetchSpT()` and `FetchSpTAK()` for species, and `FetchSpT()` and `FetchSpTAK()` for genus, extract citation records for index calculations. These functions are further distinguished by their suffixes "T" and "TAK". "T" functions only find publications with the species' name in the publication title, whereas "TAK" functions query articles' title, abstract and keywords. Apart from the *h*-index, `specieshindex` can also compute other established influence indices, including the *h5* index, the *m*-index, and the *i10* index. The *h5* index is the *h*-index of articles published in the most recent 5 years [@crotty2017other]. The *m*-index is the *h*-index divided by the number of years since the first publication [@hirsch_index_2005]. The *i10* index is the total number of articles with 10 or more citations; it is currently used by Google Scholar [@noruzi2016impact].
+There are two types of functions that connect to the literature databases. The count functions, e.g. `CountSpT()` and `CountSpTAK()` for species, and `CountGenusT()` and `CountGenusTAK()` for genus, return the total publication count without extracting any citation records. Whereas the fetch functions, e.g. `FetchSpT()` and `FetchSpTAK()` for species, and `FetchSpT()` and `FetchSpTAK()` for genus, extract citation records for index calculations. These functions are further distinguished by their suffixes "T" and "TAK". "T" functions only find publications with the species' name in the publication title, whereas "TAK" functions query articles' title, abstract and keywords. Apart from the *h*-index, `specieshindex` can also compute other established influence indices, including the *h5* index, the *m*-index, and the *i10* index. The *h5* index is the *h*-index of articles published in the most recent 5 years [@crotty2017other]. The *m*-index is the *h*-index divided by the number of years since the first publication [@hirsch_index_2005]. The *i10* index is the total number of articles with 10 or more citations; it is currently used by Google Scholar [@jacso2021].
 
 # Implementation
 
