+++ paper.bib
@@ -325,16 +325,17 @@
   number={35},
   pages={2646--2647},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/eurheartj/ehx446}
 }
 
-@article{noruzi2016impact,
-  title={Impact Factor, h-index, i10-index and i20-index of Webology},
-  author={Noruzi, Alireza},
-  journal={Webology},
-  volume={13},
-  number={1},
-  pages={1--4},
-  year={2016},
-  publisher={Webology Center}
+@article{jacso2021,
+  author = {Jacso, Peter},
+  year = {2012},
+  month = {02},
+  pages = {126-141},
+  title = {Google Scholar Author Citation Tracker: Is it too little, too late?},
+  volume = {36},
+  journal = {Online Information Review},
+  doi = {10.1108/14684521211209581}
 }
