### 2021-09-24 #3744 @Bisaloo
Hi @kamapu :wave:, I saw your post on rOpenSci slack and this package seems relevant to your interests. Would you be able to help JOSS with a review?
