### 2021-09-24 #3744 @kamapu
> Hi @kamapu wave, I saw your post on rOpenSci slack and this package seems relevant to your interests. Would you be able to help JOSS with a review?

This seems to be an interesting package. I may not feel comfortable reviewing a package but if the matter of review is a manuscript, I can take the challenge :+1:
