### 2021-09-24 #3744 @Bisaloo
@kamapu, thanks for your answer. You can learn more about the review process at JOSS in [our documentation](https://joss.readthedocs.io/en/latest/index.html). The section ["review criteria"](https://joss.readthedocs.io/en/latest/review_criteria.html) and ["review checklist"](https://joss.readthedocs.io/en/latest/review_checklist.html) should answer your questions.

In short, we review both [the package](https://github.com/jessicatytam/specieshindex) & [the paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03744/joss.03744/10.21105.joss.03744.pdf). I believe it is possible to bring valuable insights by trying to use the package a way any user would: try to install it, read the documentation, see if you get any errors at any point, tell us if any crucial functionality seems to be missing, etc. Some items may require a slightly more advanced knowledge of R (e.g., does the package have tests?) but seeing that you already developed a couple of packages, you might already be familiar with these concepts.

Feel free to take a couple of days to think about it if you'd like. But if you are already sure you do not wish to review this package & paper, I understand and respect your decision :+1:.
