### 2021-09-28 #3744 @Bisaloo
Hi @tom-gu :wave:, would you have any interest in helping us to review this package and paper? I understand this is only tangentially related to the core of your research interests (which I discovered via the `bdverse`) but I believe this has the potential to pique your interest :wink:
