### 2021-09-29 #3744 @Bisaloo
Hi @hope-data-science, would you be able to help us review this [R package](https://github.com/jessicatytam/specieshindex) and [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03744/joss.03744/10.21105.joss.03744.pdf) for JOSS? According to the keywords you submitted in our review volunteering form (`scientometrics`, `bibliometrics`, `ecology`), this submission is made for you :grin:
