### 2021-09-29 #3744 @tom-gu
> Hi @tom-gu 👋, would you have any interest in helping us to review this package and paper? I understand this is only tangentially related to the core of your research interests (which I discovered via the `bdverse`) but I believe this has the potential to pique your interest 😉

Hi @Bisaloo 😃
Yes, I can review the [R package](https://github.com/jessicatytam/specieshindex) and [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03744/joss.03744/10.21105.joss.03744.pdf) for JOSS.
What's a reasonable timetable?
