### 2021-09-29 #3744 @Bisaloo
Thanks for your answer! The review at JOSS is slightly different than others journals. In short, we have found it is usually more efficient to have a more interactive review, where reviewers can post their comments/suggestions/questions as they come as GitHub issues in the submission repository. This differs from "traditional" reviews where you send all your review comments at once. [More details are available in our documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I'll post a detailed guide when the review process officially starts.

We usually aim to have comments within 1-2 weeks (you will receive an automated reminder after 2 weeks asking you about your current progress) but there is room for some flexibility.
