### 2021-09-29 #3744 @Bisaloo
> > Hi @DrMattG wave, would you be able to help us review this [R package](https://github.com/jessicatytam/specieshindex) and [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03744/joss.03744/10.21105.joss.03744.pdf) for JOSS? This seems _really_ close to the interests you list in our review volunteering form and in your [GitHub README](https://github.com/DrMattG/DrMattG)!
>
> Hi @Bisaloo I'd be happy to!

Amazing! With 3 reviewers, we are ready to go! The review can now formally start :tada:.
