### 2021-09-29 #3776 @whedon
[REVIEW]: specieshindex: How scientifically popular is a species?
**Submitting author:** @jesstytam (<a href="http://orcid.org/0000-0003-3655-1974">Jessica Tam</a>)
**Repository:** https://github.com/jessicatytam/specieshindex
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.1

- [x] #3902
