### 2021-09-29 jessicatytam/specieshindex#44 @Bisaloo
Missing DOI in JOSS paper
Hi,

I found a missing DOI that whedon didn't spot in https://github.com/openjournals/joss-reviews/issues/3776:

- [10.1016/S1697-2600(14)70050-X](https://doi.org/10.1016/S1697-2600(14)70050-X) for 'The meaning of the h-index'
