+++ paper.bib
@@ -22,6 +22,7 @@
   pages={161--164},
   year={2014},
   publisher={Elsevier}
+  doi={10.1016/S1697-2600(14)70050-X}
 }
 
 @article{cox_comparison_2016,
