### 2021-09-30 17f0619
updated sp_check()
