### 2021-10-19 jessicatytam/specieshindex#46 @DrMattG
[JOSS REVIEW] Functionality - error with Fetch functions Bad Request (HTTP 400) 
[Review = https://github.com/openjournals/joss-reviews/issues/3776]
The Fetch functions (db="scopus") give an Bad Request (HTTP 400) error.

`library(specieshindex)

CountSpT(db = "scopus", genus = "Bettongia", species = "penicillata")

#> [1] "Species found on CoL, ITIS, NCBI, or EoL."
#> [1] 45

FetchSpT(db = "scopus", genus = "Bettongia", species = "penicillata")

#> [1] "Species found on CoL, ITIS, NCBI, or EoL."
#> [1] "45 records found."
#> [1] "starting iteration:  0  Note: iteration size is  1000  records, which runs of 200 records inside each iteration."
#> [1] "Fetching records now."
#> The query list is:
#> list(query = "TITLE(\"Bettongia penicillata\")", count = 200,
#>     start = 0, view = "STANDARD")
#> $query
#> [1] "TITLE(\"Bettongia penicillata\")"
#>
#> $count
#> [1] 200
#>
#> $start
#> [1] 0
#>
#> $view
#> [1] "STANDARD"
#>
#> Response [https://api.elsevier.com/content/search/scopus?query=TITLE%28%22Bettongia%20penicillata%22%29&count=200&start=0&view=STANDARD]
#>   Date: 2021-10-19 07:25
#>   Status: 400
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 131 B
#> Error in get_results(query, start = init_start, count = count, verbose = verbose, : Bad Request (HTTP 400).

CountGenusT(db = "scopus", genus = "Bettongia")

#> [1] "Genus found on CoL, ITIS, NCBI, or EoL."
#> [1] 122

FetchGenusT(db = "scopus", genus = "Bettongia")

#> [1] "Genus found on CoL, ITIS, NCBI, or EoL."
#> [1] "122 records found."
#> [1] "starting iteration:  0  Note: iteration size is  1000  records, which runs of 200 records inside each iteration."
#> [1] "Fetching records now."
#> The query list is:
#> list(query = "TITLE(\"Bettongia\")", count = 200, start = 0,
#>     view = "STANDARD")
#> $query
#> [1] "TITLE(\"Bettongia\")"
#>
#> $count
#> [1] 200
#>
#> $start
#> [1] 0
#>
#> $view
#> [1] "STANDARD"
#>
#> Response [https://api.elsevier.com/content/search/scopus?query=TITLE%28%22Bettongia%22%29&count=200&start=0&view=STANDARD]
#>   Date: 2021-10-19 07:25
#>   Status: 400
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 131 B
#> Error in get_results(query, start = init_start, count = count, verbose = verbose, : Bad Request (HTTP 400).
#`
