### 2021-10-19 jessicatytam/specieshindex#47 @DrMattG
[JOSS REVIEW] Add community guidlines 
[Review = https://github.com/openjournals/joss-reviews/issues/3776]
It is really helpful to add some guidelines on how to contribute to the package (this includes issues - bug reports & enhancements as well as pull requests)
