### 2021-10-19 jessicatytam/specieshindex#48 @DrMattG
[JOSS REVIEW] - Software paper bibtex issues
[Review = https://github.com/openjournals/joss-reviews/issues/3776]

There are a couple of very minor issues with the paper - in the bibtex file you are missing a comma (I will add a pull request for this). Also in the bibtex you have a paper by "Trimble & van Aarde" - but this renders as "Trimble & Aarde". Prof van Aarde was my PhD supervisor and he would not be happy to lose part of his surname ;) I could not work out how to fix this but there must be a way.
