+++ paper.bib
@@ -21,7 +21,7 @@
   number={2},
   pages={161--164},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/S1697-2600(14)70050-X}
 }
 
@@ -297,7 +297,7 @@
 }
 
 @article{trimble_aarde2010,
-  author = {Trimble, Morgan and van Aarde, Rudi},
+  author = {Trimble, Morgan and {van Aarde}, Rudi},
   year = {2010},
   month = {02},
   pages = {886-90},
