### 2021-10-21 3419317
fixing FetchGenus functions
