+++ specieshindex.Rmd
@@ -957,9 +957,9 @@ FetchSpTAK <- function(db, genus, species, synonyms, additionalkeywords, languag
 #fetchgenust() wrapper
 FetchGenusT <- function(db, genus, synonyms, additionalkeywords, language = 0) {
   if (db == "scopus") {
-    fetchsp <- FetchGenusT_scopus(genus, synonyms, additionalkeywords, language)
+    fetchgenus <- FetchGenusT_scopus(genus, synonyms, additionalkeywords, language)
   } else if (db == "wos") {
-    fetchsp <- FetchGenusT_wos(genus, synonyms, additionalkeywords)
+    fetchgenus <- FetchGenusT_wos(genus, synonyms, additionalkeywords)
   }
   return(fetchgenus)
 }
@@ -969,9 +969,9 @@ FetchGenusT <- function(db, genus, synonyms, additionalkeywords, language = 0) {
 #fetchgenustak() wrapper
 FetchGenusTAK <- function(db, genus, synonyms, additionalkeywords, language = 0) {
   if (db == "scopus") {
-    fetchsp <- FetchGenusTAK_scopus(genus, synonyms, additionalkeywords, language)
+    fetchgenus <- FetchGenusTAK_scopus(genus, synonyms, additionalkeywords, language)
   } else if (db == "wos") {
-    fetchsp <- FetchGenusTAK_wos(genus, synonyms, additionalkeywords)
+    fetchgenus <- FetchGenusTAK_wos(genus, synonyms, additionalkeywords)
   }
   return(fetchgenus)
 }
@@ -2396,6 +2396,7 @@ CountSpT("scopus", "macropus", "rufus") #114
 CountSpT("wos", "macropus", "rufus") #113
 CountSpT("base", "macropus", "rufus") #152
 CountSpT(db = "base", genus = "Bettongia", species = "penicillata")
+FetchGenusT(db = "scopus", genus = "Bettongia")
 
 
 genus <- "Bettongia"
