### 2021-10-21 jessicatytam/specieshindex#46 @jessicatytam
Seems to be working now, I've forgotten to replace `sp` with `genus` when copying things over to `FetchGenusT()` and `FetchGenusTAK()`. Do let me know if that has fixed things.
