### 2021-10-21 316e7a7
trying to fix github actions issue based on
https://github.community/t/hosted-runners-not-picking-up-jobs/191663/13
