### 2021-10-21 c13a6b3
added community guidelines to README
