+++ specieshindex.Rmd
@@ -2292,103 +2292,6 @@ Spi10(dummy1) #12
 apikey <- "442b9048417ef20cf680a0ae26ee4d86"
 sid <- auth(username = NULL, password = NULL)
 
-CountSpT <- function(db, genus, species, synonyms, additionalkeywords, size = 50000) {
-  if (db == "scopus") {
-    countsp <- CountSpT_scopus(genus, species, synonyms, additionalkeywords)
-  } else if (db == "wos") {
-    countsp <- CountSpT_wos(genus, species, synonyms, additionalkeywords)
-  } else if (db == "base") {
-    countsp <- CountSpT_base(genus, species, synonyms, additionalkeywords)
-  } else if (db == "lens") {
-    countsp <- CountSpT_lens(genus, species, synonyms, additionalkeywords, size)
-  }
-  return(countsp)
-}
-
-CountSpT_wrap("scopus", genus = "Osphranter", species = "rufus") #4
-CountSpT_wrap("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus") #118
-CountSpT_wrap("scopus", genus = "Osphranter", species = "rufus", additionalkeywords = "conserv*") #0
-CountSpT_wrap("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*") #1
-
-CountSpT_wrap("wos", genus = "Osphranter", species = "rufus") #3
-CountSpT_wrap("wos", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus") #116
-CountSpT_wrap("wos", genus = "Osphranter", species = "rufus", additionalkeywords = "conserv*") #0
-CountSpT_wrap("wos", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*") #1
-
-test <- FetchSpT("scopus", genus = "Osphranter", species = "rufus", language = 1)
-
-
-
-load("data/languages.rda")
-write.csv(languages, "data/languages.csv")
-
-#ggplot
-
-test_bind <- list(Koala, Platypus, Quokka, Woylie)
-
-test_bind <- test_bind %>% lapply(test_bind, function(x) mutate(sp = as.character(substitute(x))))
-
-
-plotPub(Koala)
- 
-
-#genus
-
-create_query_string_T_scopus_genus <- function(genus, synonyms, additionalkeywords){
-  if (missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE("', genus, '")'))
-  } 
-  if (!missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE("', genus, '" AND ', additionalkeywords, ')'))
-  }
-  if (missing(additionalkeywords) & !missing(synonyms)) {
-    temp_string <- paste0('TITLE("', genus, '" OR "', synonyms[1], '"')
-    if (length(synonyms)==1) {
-      return(paste0(temp_string, ')'))
-    }
-    else {
-      for (i in 2:length(synonyms)){
-        temp_string <- paste0(temp_string, ' OR "', synonyms[i], '"')
-      }
-      return(paste0(temp_string, ')'))
-    }
-  }
-  if (!missing(additionalkeywords) & !missing(synonyms)) {
-    return(paste0('TITLE(("', genus, '" OR "', synonyms, '") AND ', additionalkeywords, ')'))
-  } 
-}
-
-
-CountSpT_scopus_genus <- function(genus, synonyms, additionalkeywords, datatype = "application/xml") {
-  genus_check(genus)
-  response <- httr::GET("http://api.elsevier.com/content/search/scopus",
-                      query = list(apiKey = apikey,
-                                   query = create_query_string_T_scopus_genus(genus, synonyms, additionalkeywords),
[…]
