### 2021-10-28 2b781d3
trying both XML and JSON for BASE functions
