### 2021-11-05 jessicatytam/specieshindex#51 @Bisaloo
Add tests
The package currently has some tests but the [coverage](https://en.wikipedia.org/wiki/Code_coverage) is very limited since tests are only checking the built-in data and none of the actual features of the package.

Before acceptance in JOSS (https://github.com/openjournals/joss-reviews/issues/3776), it will probably be necessary to increase the number of tests and the portion of the code covered by tests.

Among other things, the bugs uncovered by the reviewers would be good examples of things to add in the tests.

The fact that you already have some tests suggest you know how to do this but regarding the specific topic of testing HTTP resources, you might find the book ['HTTP testing in R' by Maëlle Salmon and Scott Chamberlain](https://books.ropensci.org/http-testing/) useful. Please let me know if you'd like more guidance on this topic.
