+++ tests/testthat/test_woylie.R
@@ -12,7 +12,7 @@ test_that("Woylie_stats", {
   expect_true(Spi10(Woylie)>=54)
   expect_true(SpH5(Woylie)>=5)
   expect_true(SpHAfterdate(Woylie, "2000-01-01")>=20)
-  B_penicillata <- Allindices(Woylie,"Bettongia","penicillata")
+  B_penicillata <- Allindices(Woylie, "Bettongia", "penicillata")
   expect_true(is.data.frame(B_penicillata))
 })
 #
