+++ tests/testthat/test_plots.R
@@ -10,6 +10,9 @@ test_that("preparing_data_for_plots", {
   expect_true(is.data.frame(B_penicillata))
 })
 
-# test_that("plotting", {
-#   
-# })
+test_that("plotting", {
+  expect_identical(class(index_plot$layers[[1]]$geom)[1], "GeomDrawGrob")
+  expect_identical(class(index_plot$layers[[1]]$stat)[1], "StatIdentity")
+  expect_identical(class(pub_plot$layers[[1]]$geom)[1], "GeomLine")
+  expect_identical(class(pub_plot$layers[[1]]$stat)[1], "StatIdentity")
+})
