+++ .github/workflows/main.yml
@@ -40,6 +40,15 @@ jobs:
           r-version: ${{ matrix.config.r }}
 
       - uses: r-lib/actions/setup-pandoc@v1
+      
+      - uses: actions/checkout@master
+      
+      - uses: codecov/codecov-action@v2
+        with:
+          files: ./coverage1.xml,./coverage2.xml # optional
+          flags: unittests # optional
+          name: codecov-umbrella # optional
+          verbose: true # optional (default = false)
 
       - name: Query dependencies
         run: |
@@ -82,3 +91,5 @@ jobs:
         with:
           name: ${{ runner.os }}-r${{ matrix.config.r }}-results
           path: check
+          
+          
