+++ README.rmd
@@ -8,7 +8,7 @@ output: github_document
 
 [![CRAN status](https://www.r-pkg.org/badges/version/specieshindex)](https://CRAN.R-project.org/package=specieshindex)
 ![R-CMD-check](https://github.com/jessicatytam/specieshindex/workflows/CI/badge.svg)
-`r badger::badge_codecov("jessicatytam/specieshindex")`
+[![codecov](https://codecov.io/gh/jessicatytam/specieshindex/branch/master/graph/badge.svg?token=Y8N1QW0I1C)](https://codecov.io/gh/jessicatytam/specieshindex)
 [![Github All Releases](https://img.shields.io/github/downloads/jessicatytam/specieshindex/total.svg)]()
 
 `specieshindex` is a package that aims to gauge scientific influence of different species mainly using the *h*-index.
