+++ tests/testthat/test_auth.R
@@ -1,0 +1,6 @@
+context("testing_authentication")
+
+test_that("wos_sid_presence", {
+  sid <<- wosr::auth()
+  expect_true(is.character(sid))
+})
