### 2021-11-10 #3776 @kamapu
> @kamapu, you should be able to access the manuscript via the links in the previous comment slightly_smiling_face

Thanks! The robot is closing my issues in the repo of specieshindex. I'm doing something wrong or it is because of been too late.
