### 2021-11-10 #3776 @tom-gu
@jessicatytam @wcornwell @Bisaloo
Here is my full review, **so sorry for the delay**!

----

## Package overview:
The `specieshindex` package is showing great potential to promote scientific merit. It is a delightfully user-friendly package with superb documentation. Nonetheless, I identified several areas that can be improved. I did not open specific GitHub issues since they are more systemic and should be addressed as the package matures. Additionally, I had a few minor suggestions and some ideas for future development.

## Package structure & development:

### Testing & CI
I saw that you recently added several tests and a CI (via GithubAction). It is great to know that you are adopting a more test-driven development (TDD) schema. Pivotal tests for this particular package are smoke tests to monitor API connection and to validate a proper output structure. As you know by now, constructing this for [Scopus ](https://github.com/jessicatytam/specieshindex/blob/master/tests/testthat/test_scopus.R) is easier than for [WOS](https://github.com/jessicatytam/specieshindex/blob/master/tests/testthat/test_auth.R) and BASE. However, please refer to [this section](https://devguide.ropensci.org/package-development-security-best-practices.html#pkgsecrets) to learn how to manage secrets (e.g., API keys) within the package.

### Code duplication
The level of duplication (code & function documentation) is off the charts. It partially stems from a segregated function strategy and from not fully utilizing sub-functions (internal, non-exported functions). This is problematic from the POV of a package developer since even the smallest changes can become a real hassle. When the small things become exhausting, the big things become impossible!
I’m advocating for the consolidation of similar functions by adding a parameter or two. This might slightly affect the user’s experience, but it will greatly serve package maintenance and improve its overall structure. If you decide to make the changes, please avoid putting all the functions in the same R file.

### Code refactoring (mostly coding efficiency)
Going forward, I encourage you to review your coding efficiency. Few examples:
- Is it necessary to [hit the API so many times](https://github.com/jessicatytam/specieshindex/blob/9698952bf81d8d549b7c5bef3a4a557f342bf13d/R/specieshindex.R#L472)? Can post-processing of one API call achieve the same results? What will happen when the current year is 2022 or 2023? Isn’t the raw data by itself valuable?
- [When plotting all indices](https://github.com/jessicatytam/specieshindex/blob/9698952bf81d8d549b7c5bef3a4a557f342bf13d/R/specieshindex.R#L1991), why not using facet_wrap() or facet_grid() for multi-panel plots?

### Defensive programming
Communicating with our users via errors, warnings, and messages significantly improves their user experience and can help mitigate knowledge gaps. Currently, the package has no implementation of defensive programming. I encourage you to embrace it. [Here you can helpful information](https://adv-r.hadley.nz/conditions.html#signalling-conditions). Side note: implementing defensive programming is much harder when you have high levels of code duplication.

### Motivation for package maintenance
Package maintenance is crucial for establishing software sustainability. This aspect of package maintenance is a demanding God. Adequately quantifying long-term maintenance is rather challenging, even more so for scientists. When trying to connect the dots, it seems that this package was mainly developed for the study of [biases in mammalian](https://github.com/jessicatytam/biases_in_mammalian_research). If this is the case, I wonder if the Cornwell Lab has a long-term maintenance plan with sufficient resources to support it. Please forgive me if the query seems intrusive, I’m only trying to share insights from my own experience and pain. It might be beneficial to instigate a collaboration with an organization or additional PI that is going to heavily rely on this package. Perhaps @mjwestgate will have specific leads.

### Data science >> Software development
I can classify this package under ‘data scientists developing a package’ scenario. It’s clear that you completely understand your user, but do not yet fully comprehend your development needs. In order to support your package development journey, I urge you to read [rOpenSci Packaging Guide](https://devguide.ropensci.org/index.html) and [R Packages](https://r-pkgs.org/). In a way, now is the perfect time for that. Now, you properly digest these comprehensive resources.

## goodpractice output
See: https://github.com/MangoTheCat/goodpractice
Please run `goodpractice::gp()` or see [this RStudio Project](https://rstudio.cloud/project/3161383) for the goodpractice output.

## Minor suggestions:

### Installation instructions (README; paper; vignette)
Why you need all of this: https://github.com/jessicatytam/specieshindex/blob/9698952bf81d8d549b7c5bef3a4a557f342bf13d/README.rmd#L22:L31
It’s somewhat redundant, why not just:
```
install.packages("remotes")
remotes::install_github("jessicatytam/specieshindex")
```

### paper.md file
Code is too long in lines 76, 109, 144-147, and 150.
https://raw.githubusercontent.com/openjournals/joss-papers/joss.03776/joss.03776/10.21105.joss.03776.pdf

### DESCRIPTION file
Package contributors might be missing (compared to [paper authors](http://authors)).

### git
I see that both Jessica and Will are committing directly to the master branch. After pkg release, it’s better to have a dev branch and to merge code via PR.

### pkgdown
The overall level of package documentation is excellent. Creating a website via [`pkgdown`](https://pkgdown.r-lib.org/) will beautifully showcase it. Implementation is quite easy and totally worth it!

## Ideas for future functionality:

- Introducing a query ‘batch mode’ by passing a dataframe
- To utilize taxize to transparently generate the list of synonyms

## Final words
Sharing as many insights as I can is not just my duty but also a privilege, as it makes my own hardships along the way even more worthwhile. I hope you will find this review helpful and not too preachy 😃

From one Biodiversity Jedi to another, may the foRce be with you!
Tomer
