+++ README.rmd
@@ -29,7 +29,7 @@ install.packages("httr")
 install.packages("dplyr")
 install.packages("data.table")
 install.packages("tidyr")
-devtools::install_github("jessicatytam/specieshindex", force = TRUE, build_vignettes = FALSE)
+devtools::install_github("jessicatytam/specieshindex")
 
 #Load the library
 library(specieshindex)
