### 2021-11-23 3d8f8cc
updated paper 'connecting to scopus' instructions
