### 2021-11-23 d62dc00
updated paper 'implementation' section
