### 2021-11-26 d11bab9
setting up scopus responses for api tests
