+++ tests/testthat/scopus-get/api.elsevier.com/content/search/scopus.R
@@ -1,35 +1,0 @@
-structure(list(url = "https://api.elsevier.com/content/search/scopus", 
-    status_code = 401L, headers = structure(list(date = "Wed, 10 Nov 2021 00:42:18 GMT", 
-        `content-type` = "application/json;charset=UTF-8", `content-length` = "112", 
-        connection = "keep-alive", allow = "GET", vary = "Origin", 
-        vary = "Access-Control-Request-Method", vary = "Access-Control-Request-Headers", 
-        `x-els-reqid` = "c40b2e91daae85fd", `x-els-resourceversion` = "default", 
-        `x-els-status` = "AUTHENTICATION_ERROR - No APIKey provided for request", 
-        `x-els-transid` = "4e6d5a0668f1cbb6", `cf-cache-status` = "DYNAMIC", 
-        `expect-ct` = "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\"", 
-        server = "cloudflare", `cf-ray` = "6abb3d7a683a6a0e-SYD"), class = c("insensitive", 
-    "list")), all_headers = list(list(status = 308L, version = "HTTP/1.1", 
-        headers = structure(list(date = "Wed, 10 Nov 2021 00:42:18 GMT", 
-            `content-type` = "text/html", `transfer-encoding` = "chunked", 
-            connection = "keep-alive", location = "https://api.elsevier.com/content/search/scopus", 
-            `cf-cache-status` = "DYNAMIC", server = "cloudflare", 
-            `cf-ray` = "6abb3d78bcfefd4e-SYD"), class = c("insensitive", 
-        "list"))), list(status = 401L, version = "HTTP/1.1", 
-        headers = structure(list(date = "Wed, 10 Nov 2021 00:42:18 GMT", 
-            `content-type` = "application/json;charset=UTF-8", 
-            `content-length` = "112", connection = "keep-alive", 
-            allow = "GET", vary = "Origin", vary = "Access-Control-Request-Method", 
-            vary = "Access-Control-Request-Headers", `x-els-reqid` = "c40b2e91daae85fd", 
-            `x-els-resourceversion` = "default", `x-els-status` = "AUTHENTICATION_ERROR - No APIKey provided for request", 
-            `x-els-transid` = "4e6d5a0668f1cbb6", `cf-cache-status` = "DYNAMIC", 
-            `expect-ct` = "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\"", 
-            server = "cloudflare", `cf-ray` = "6abb3d7a683a6a0e-SYD"), class = c("insensitive", 
-        "list")))), cookies = structure(list(domain = logical(0), 
-        flag = logical(0), path = logical(0), secure = logical(0), 
-        expiration = structure(numeric(0), class = c("POSIXct", 
-        "POSIXt")), name = logical(0), value = logical(0)), row.names = integer(0), class = "data.frame"), 
-    content = charToRaw("{\"service-error\":{\"status\":{\"statusCode\":\"AUTHENTICATION_ERROR\",\"statusText\":\"No APIKey provided for request\"}}}"), 
-    date = structure(1636504938, class = c("POSIXct", "POSIXt"
-    ), tzone = "GMT"), times = c(redirect = 0.25813, namelookup = 0.008264, 
-    connect = 0.010096, pretransfer = 0.022962, starttransfer = 0.340852, 
-    total = 0.599018)), class = "response")
