### 2021-11-26 5d12b4c
added tests for CountSpT_scopus
