### 2021-11-28 #3776 @kamapu
# Review by Miguel

Sorry for the late response. It is my first review of a package and the way to proceed has been challenging for me. Also some "spontaneous" activities were kipping me busy and I just managed to finish my review.

While this package seems to be promising for its use also in the context of my own work, I see the necessity of a major revision of both, the structure of functions as well as in the documentation.

I didn't opened issues for this review. I would suggest to the authors when responding to refer every change to a commit in the repository.

I sum up my suggestions here (detailed comments below):

- Implement a secure way to provide credentials in sessions.
- Re-structure the documentation for the  functions `Count*()` and `Fetch*()`. In the best case, define just two functions with all possible variants set through the arguments.
- Examples in documentation and tutorials should follow the whole path from querying to assessing, discarding the use of pre-stored data sets.
- Get rid of the pre-stored data. The user have to be able, first of repeat the query and get the results or apply the queries to own questions. Pre-stored data can still be hidden and used for making vignettes and avoid possible errors but giving examples to the users on the basis of such pre-stored data is in my opinion rather confusing.
- Not mandatory, but a function querying the metadata of retrieved publications (e.g. for its export to **BibTeX**) will complete the whole story about `specieshindex`.

## Comments and Suggestions

- [ ] I also support the idea of simplifying code, for instance using following code (include the same in the vignette):
`remotes::install_github("jessicatytam/specieshindex", build_vignettes = TRUE, dependencies = TRUE)`
- [ ] In relation to the last, I was wondering why the the suggested installation from github is not including `build_vignettes =  RUE`. If you are fearing, the users may not have any LaTeX distribution, you can rather use HTML as output format. Users may still need to install Rtools and pandoc, though.

### Vignette

- [ ] Call the vignette with a topic name, not just `vignette`.
- [ ] Instructions for installing (same comment as for README).
- [ ] In chapter **Workflow** a brief description of the databases is required, especially regarding BASE (I didn't knew this one).
- [ ] In chapter **Other Arguments and Syntax** you should invert the sorting of the sub-chapters **Additional Keywords** and  *Count, Fetch, ...** because you explain the functions and their difference in the second one and should be done at the  beginning to avoid confusions. For instance I didn't noticed that there is a block using `CountSpT()` and `CountSpTAK()`.
- [ ] in `FetchSpT(db = "scopus", genus = "Bettongia", species = "penicillata")` I get an error:
```
Error in get_results(query, start = init_start, count = count, verbose = verbose,  :
  Bad Request (HTTP 400).
```
- [ ] I'm wondering why the results of `FetchSp*()` or `CountSp*()` cannot be assigned to an R object. This is very restrictive in the case that you like to implement these functions in  statistical assessments or reports.

### Manual

- [ ] My impression is that the manual can be arranged in a better way and perhaps even the functions can be simplified. For instance all `FetchSP*()` functions should share the same documentation page. The same applies for `CountSP*()`.
In fact, I would even be more radical and just define two function `count()` and `fetch()` (names just as examples), where everything else (either looking for species or genera, in title, in keywords, in abstract, in which database, etc.) can be set as argument in ad-hoc function parameters.
- [ ] I failed to assign outputs of `Count*()` and `Fetch*()` to an object (e.g. by `obj <- Fetch*()`). I also wonder that any of the examples do assignments in the manual and all functions that are supposed to work on outputs of these functions actually work on data already installed by the package. I know that such kind of queries may fail when building the package but you have the option of writing your examples within a `\dontrun{}` environment. I will prefer to provide examples to the users starting by the query.
- [ ] If addressed all previous issues, installed data will become superfluous.
- [ ] Is there an option for extract the metadata of the queried references, for
instance to export it to a **BibTeX** file?

### Manuscript

- [ ] The same comments for Vignette and Manual may be also applied to the manuscript.
- [ ] I would prefer to have a concrete example of a question and apply it from the query to reporting the outcome without using pre-stored data.

- [ ] I found quite risky the use of objects for the storage of credentials because by sharing a project you may mistakenly share a script or an object revealing your passwords and IDs. An alternative can be to use masks with `tcltk` and store the entries in the temporary folder (see [this post](http://rstat.consulting/blog/temporary-dir-and-files-in-r/)) or using the package `keyring` (see [here](https://datacornering.com/how-to-use-password-and-token-securely-in-r/)).

## Checks

- [ ] Current checks retrieves warnings about the use of `:::` and `::` which may be avoided. The second case, as far as I know, may overcome the declaration of dependencies, which is not desirable.
- [ ] Warnings regarding `no visible binding for global variable ...`. In this case I'll recommend to use **NULLing** (see [this example](https://github.com/kamapu/vegtable/blob/master/R/NULLing.R)).
- [ ] Check and solve all issues retrieved by `goodpractice::gp()`
