### 2021-12-03 jessicatytam/specieshindex#46 @jessicatytam
I think your institution might not be a Scopus/Elsevier paid subscriber, which means they don't allow paper extraction but you can still count the number of papers.
