### 2021-12-06 c170221
addressing comments from @kamapu re:README, vignette, and MS
