+++ paper.bib
@@ -340,3 +340,13 @@
   journal = {Online Information Review},
   doi = {10.1108/14684521211209581}
 }
+
+@misc{tam_lagisz_cornwell_nakagawa_2021,
+ title={Quantifying research interests in 7,521 mammalian species with h-index: a case study},
+ url={ecoevorxiv.org/gd7cv},
+ DOI={10.32942/osf.io/gd7cv},
+ publisher={EcoEvoRxiv},
+ author={Tam, Jessica and Lagisz, Malgorzata and Cornwell, William K and Nakagawa, Shinichi},
+ year={2021},
+ month={Nov}
+}
