+++ vignettes/references.bib
@@ -149,3 +149,13 @@ source={Scopus},
 	eprint = {https://doi.org/10.3897/zookeys.550.9546},
 	journal = {ZooKeys}
 }
+
+@misc{tam_lagisz_cornwell_nakagawa_2021,
+ title={Quantifying research interests in 7,521 mammalian species with h-index: a case study},
+ url={ecoevorxiv.org/gd7cv},
+ DOI={10.32942/osf.io/gd7cv},
+ publisher={EcoEvoRxiv},
+ author={Tam, Jessica and Lagisz, Malgorzata and Cornwell, William K and Nakagawa, Shinichi},
+ year={2021},
+ month={Nov}
+}
