### 2021-12-07 354c07e
fixing tests for scopus requests
