+++ R/specieshindex.R
@@ -287,6 +287,9 @@ FetchGenusTAK <- function(db, genus, synonyms, additionalkeywords, language = 0)
 #' 
 #' CountSpT_scopus("bettongia", "penicillata", "conserv*")
 #' }
+#' 
+#' @noRd
+#' 
 CountSpT_scopus <- function(genus, species, synonyms, additionalkeywords, datatype = "application/xml") {
   sp_check(genus, species)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
@@ -334,6 +337,9 @@ CountSpT_scopus <- function(genus, species, synonyms, additionalkeywords, dataty
 #' 
 #' CountSpTAK_scopus("bettongia", "penicillata")
 #' }
+#' 
+#' @noRd
+#' 
 CountSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, datatype = "application/xml") {
   sp_check(genus, species)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
@@ -380,6 +386,9 @@ CountSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, data
 #' 
 #' CountGenusT_scopus("bettongia", "conserv*")
 #' }
+#' 
+#' @noRd
+#' 
 CountGenusT_scopus <- function(genus, synonyms, additionalkeywords, datatype = "application/xml") {
   genus_check(genus)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
@@ -426,6 +435,9 @@ CountGenusT_scopus <- function(genus, synonyms, additionalkeywords, datatype = "
 #' 
 #' CountGenusTAK_scopus("bettongia")
 #' }
+#' 
+#' @noRd
+#' 
 CountGenusTAK_scopus <- function(genus, synonyms, additionalkeywords, datatype = "application/xml") {
   genus_check(genus)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
@@ -469,6 +481,9 @@ CountGenusTAK_scopus <- function(genus, synonyms, additionalkeywords, datatype =
 #' 
 #' FetchSpT_scopus("bettongia", "penicillata", "conserv*")
 #' }
+#' 
+#' @noRd
+#' 
 FetchSpT_scopus <- function(genus, species, synonyms, additionalkeywords, language = 0) {
   count <- CountSpT_scopus(genus, species, synonyms, additionalkeywords) #check the number of records
   print(paste(count, "records found."))
@@ -639,6 +654,9 @@ FetchSpT_scopus <- function(genus, species, synonyms, additionalkeywords, langua
 #' 
 #' FetchSpTAK_scopus("bettongia", "penicillata", "conserv*")
 #' }
+#' 
+#' @noRd
+#' 
 FetchSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, language = 0) {
   count <- CountSpTAK_scopus(genus, species, synonyms, additionalkeywords) #check the number of records
   print(paste(count, "records found."))
@@ -808,6 +826,9 @@ FetchSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, lang
 #' 
 #' FetchGenusT_scopus("bettongia", "conserv*")
 #' }
+#' 
+#' @noRd
+#' 
 FetchGenusT_scopus <- function(genus, synonyms, additionalkeywords, language = 0) {
   count <- CountGenusT_scopus(genus, synonyms, additionalkeywords) #check the number of records
   print(paste(count, "records found."))
[…]
