+++ man/FetchGenusTAK_wos.Rd
@@ -1,38 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/specieshindex.R
-\name{FetchGenusTAK_wos}
-\alias{FetchGenusTAK_wos}
-\title{Fetch data from Web of Science - title, abstract and author keywords.}
-\usage{
-FetchGenusTAK_wos(genus, synonyms, additionalkeywords)
-}
-\arguments{
-\item{genus}{Genus classification from the binomial name.}
-
-\item{synonyms}{Alternate genus names.}
-
-\item{additionalkeywords}{Optional search terms.}
-}
-\value{
-A dataframe of the genus' citation records with the given \code{genus}.
-}
-\description{
-This function fetches citation information from Web of Science using genus name found in the title, abstract and author keywords of the publications.
-Duplicates are to be removed by the user after fetching the data.
-}
-\examples{
-\dontrun{
-FetchGenusTAK_wos("Bettongia")
-
-#lower case letter in genus is also accepted and will return identical results
-
-FetchGenusTAK_wos("bettongia")
-}
-\dontrun{
-FetchGenusTAK_wos("Bettongia", "conserv*")
-
-#lower case letter in genus is also accepted and will return identical results
-
-FetchGenusTAK_wos("bettongia", "conserv*")
-}
-}
