### 2021-12-08 03d0f8b
updating documentation and fixing checks
