+++ R/specieshindex.R
@@ -1,255 +1,131 @@
-#' This is a wrapper function for \code{\link{CountSpT_scopus}}, \code{\link{CountSpT_wos}}, and \code{\link{CountSpT_base}}.
+#' This function counts the total number of search results from Scopus, Web of Science, or BASE.
+#' A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
 #' 
-#' @title CountSpT wrapper
+#' @title Search count of literature
 #'
 #' @param db Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").
+#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
+#' @param level Taxonomic level. Genus ("genus") or species ("species").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
 #' @param synonyms Alternate species names.
 #' @param additionalkeywords Optional search terms.
-#' 
-#' @return Search count of the species with the given \code{genus} and \code{species}.
+#'
+#' @return Search count of the genus or species with the given \code{genus} and/or \code{species}.
 #' @export
 #'
 #' @examples
 #' \dontrun{
-#' CountSpT("scopus", genus = "Osphranter", species = "rufus")
+#' Count(db = "scopus",
+#'       search = "t",
+#'       level = "species",
+#'       genus = "Osphranter", species = "rufus")
 #' }
 #' \dontrun{
-#' CountSpT("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*")
+#' Count(db = "scopus",
+#'       search = "t",
+#'       level = "species",
+#'       genus = "Osphranter", species = "rufus",
+#'       synonyms = "Macropus rufus",
+#'       additionalkeywords = "conserv*")
 #' }
-CountSpT <- function(db, genus, species, synonyms, additionalkeywords) {
-  if (db == "scopus") {
-    countsp <- CountSpT_scopus(genus, species, synonyms, additionalkeywords)
-  } else if (db == "wos") {
-    countsp <- CountSpT_wos(genus, species, synonyms, additionalkeywords)
-  } else if (db == "base") {
-    countsp <- CountSpT_base(genus, species, synonyms, additionalkeywords)
-  } 
-  return(countsp)
+Count <- function(db,
+                  search,
+                  level,
+                  genus,
+                  species,
+                  synonyms,
+                  additionalkeywords) {
+if (db == "scopus" && search == "t" && level == "species") {
+  countsp <- CountSpT_scopus(genus, species, synonyms, additionalkeywords)
+} else if (db == "wos" && search == "t" && level == "species") {
+  countsp <- CountSpT_wos(genus, species, synonyms, additionalkeywords)
+} else if (db == "base" && search == "t" && level == "species") {
+  countsp <- CountSpT_base(genus, species, synonyms, additionalkeywords)
+} else if (db == "scopus" && search == "tak" && level == "species") {
+  countsp <- CountSpTAK_scopus(genus, species, synonyms, additionalkeywords)
+} else if (db == "wos" && search == "tak" && level == "species") {
+  countsp <- CountSpTAK_wos(genus, species, synonyms, additionalkeywords)
+} else if (db == "base" && search == "tak" && level == "species") {
+  countsp <- CountSpTAK_base(genus, species, synonyms, additionalkeywords)
[…]
