+++ R/NULLing.R
@@ -1,0 +1,13 @@
+#Nulling objects
+apikey <- NULL
+sid <- NULL
+ut <- NULL
+cover_date <- NULL
+Year <- NULL
+genus_species <- NULL
+h <- NULL
+m <- NULL
+i10 <- NULL
+h5 <- NULL
+Freq <- NULL
+spp <- NULL
