+++ man/TotalPub.Rd
@@ -7,7 +7,7 @@
 TotalPub(data)
 }
 \arguments{
-\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
+\item{data}{The dataframe generated from \code{\link{Fetch}}.}
 }
 \value{
 An integer of the total number of publications.
