+++ man/TotalCite.Rd
@@ -7,7 +7,7 @@
 TotalCite(data)
 }
 \arguments{
-\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
+\item{data}{The dataframe generated from \code{\link{Fetch}}.}
 }
 \value{
 A numerical value of the total number of citations.

[…]
