### 2021-12-09 4c744f0
fixed undocumented code object warning in checks
