+++ R/specieshindex.R
@@ -36,29 +36,29 @@ Count <- function(db,
                   species,
                   synonyms,
                   additionalkeywords) {
-if (db == "scopus" && search == "t" && level == "species") {
+if (db == "scopus" & search == "t" & level == "species") {
   countsp <- CountSpT_scopus(genus, species, synonyms, additionalkeywords)
-} else if (db == "wos" && search == "t" && level == "species") {
+} else if (db == "wos" & search == "t" & level == "species") {
   countsp <- CountSpT_wos(genus, species, synonyms, additionalkeywords)
-} else if (db == "base" && search == "t" && level == "species") {
+} else if (db == "base" & search == "t" & level == "species") {
   countsp <- CountSpT_base(genus, species, synonyms, additionalkeywords)
-} else if (db == "scopus" && search == "tak" && level == "species") {
+} else if (db == "scopus" & search == "tak" & level == "species") {
   countsp <- CountSpTAK_scopus(genus, species, synonyms, additionalkeywords)
-} else if (db == "wos" && search == "tak" && level == "species") {
+} else if (db == "wos" & search == "tak" & level == "species") {
   countsp <- CountSpTAK_wos(genus, species, synonyms, additionalkeywords)
-} else if (db == "base" && search == "tak" && level == "species") {
+} else if (db == "base" & search == "tak" & level == "species") {
   countsp <- CountSpTAK_base(genus, species, synonyms, additionalkeywords)
-} else if (db == "scopus" && search == "t" && level == "genus") {
+} else if (db == "scopus" & search == "t" & level == "genus") {
   countsp <- CountGenusT_scopus(genus, synonyms, additionalkeywords)
-} else if (db == "wos" && search == "t" && level == "genus") {
+} else if (db == "wos" & search == "t" & level == "genus") {
   countsp <- CountGenusT_wos(genus, synonyms, additionalkeywords)
-} else if (db == "base" && search == "t" && level == "genus") {
+} else if (db == "base" & search == "t" & level == "genus") {
   countsp <- CountGenusT_base(genus, synonyms, additionalkeywords)
-} else if (db == "scopus" && search == "tak" && level == "genus") {
+} else if (db == "scopus" & search == "tak" & level == "genus") {
   countsp <- CountGenusTAK_scopus(genus, synonyms, additionalkeywords)
-} else if (db == "wos" && search == "tak" && level == "genus") {
+} else if (db == "wos" & search == "tak" & level == "genus") {
   countsp <- CountGenusTAK_wos(genus, synonyms, additionalkeywords)
-} else if (db == "base" && search == "tak" && level == "genus") {
+} else if (db == "base" & search == "tak" & level == "genus") {
   countsp <- CountGenusTAK_base(genus, synonyms, additionalkeywords)
 } 
 return(countsp)
@@ -106,21 +106,21 @@ Fetch <- function(db,
                   synonyms,
                   additionalkeywords,
                   language = 0) {
-  if (db == "scopus" && search == "t" && level == "species") {
+  if (db == "scopus" & search == "t" & level == "species") {
     fetchsp <- FetchSpT_scopus(genus, species, synonyms, additionalkeywords, language)
-  } else if (db == "wos" && search == "t" && level == "species") {
+  } else if (db == "wos" & search == "t" & level == "species") {
     fetchsp <- FetchSpT_wos(genus, species, synonyms, additionalkeywords)
[…]
