+++ NAMESPACE
@@ -2,27 +2,7 @@
 
 export(Allindices)
 export(Count)
-export(CountGenusTAK_base)
-export(CountGenusTAK_scopus)
-export(CountGenusTAK_wos)
-export(CountGenusT_base)
-export(CountGenusT_scopus)
-export(CountGenusT_wos)
-export(CountSpTAK_base)
-export(CountSpTAK_scopus)
-export(CountSpTAK_wos)
-export(CountSpT_base)
-export(CountSpT_scopus)
-export(CountSpT_wos)
 export(Fetch)
-export(FetchGenusTAK_scopus)
-export(FetchGenusTAK_wos)
-export(FetchGenusT_scopus)
-export(FetchGenusT_wos)
-export(FetchSpTAK_scopus)
-export(FetchSpTAK_wos)
-export(FetchSpT_scopus)
-export(FetchSpT_wos)
 export(SourceType)
 export(SpH5)
 export(SpHAfterdate)
