### 2021-12-10 39396b2
updated paper and removed dataset documentation from manual
