+++ R/Platypus.R
@@ -6,4 +6,7 @@
 #' @format A data frame with 321 rows and 20 variables
 #'
 #' @source \url{http://api.elsevier.com/content/search/scopus}
+#' 
+#' @noRd
+#' 
 "Platypus"
