+++ R/Koala.R
@@ -6,4 +6,7 @@
 #' @format A data frame with 773 rows and 20 variables
 #'
 #' @source \url{http://api.elsevier.com/content/search/scopus}
+#' 
+#' @noRd
+#' 
 "Koala"
