+++ man/Woylie.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Woylie.R
-\docType{data}
-\name{Woylie}
-\alias{Woylie}
-\title{Woylie dataset}
-\format{
-A data frame with 113 rows and 20 variables
-}
-\source{
-\url{http://api.elsevier.com/content/search/scopus}
-}
-\usage{
-Woylie
-}
-\description{
-Citation records of woylie (Bettongia penicillata) from Scopus.
-Data was retrieved on 10 July 2020.
-}
-\keyword{datasets}
