+++ man/Platypus.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Platypus.R
-\docType{data}
-\name{Platypus}
-\alias{Platypus}
-\title{Platypus dataset}
-\format{
-A data frame with 321 rows and 20 variables
-}
-\source{
-\url{http://api.elsevier.com/content/search/scopus}
-}
-\usage{
-Platypus
-}
-\description{
-Citation records of platypus (Ornithorhynchus anatinus) from Scopus.
-Data was retrieved on 10 July 2020.
-}
-\keyword{datasets}
