+++ man/Quokka.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Quokka.R
-\docType{data}
-\name{Quokka}
-\alias{Quokka}
-\title{Quokka dataset}
-\format{
-A data frame with 242 rows and 20 variables
-}
-\source{
-\url{http://api.elsevier.com/content/search/scopus}
-}
-\usage{
-Quokka
-}
-\description{
-Citation records of quokka (Setonix brachyurus) from Scopus.
-Data was retrieved on 10 July 2020.
-}
-\keyword{datasets}
