+++ R/languages.R
@@ -6,4 +6,7 @@
 #' @format A csv file with the complete list of languages of documents found on Scopus.
 #'
 #' @source \url{https://www.elsevier.com/solutions/scopus/how-scopus-works/content}
+#' 
+#' @noRd
+#' 
 "languages"
