+++ R/specieshindex.R
@@ -170,11 +170,16 @@ CountSpT_scopus <- function(genus, species, synonyms, additionalkeywords, dataty
   sp_check(genus, species)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
-                                     query = create_query_string_T_scopus(genus, species, synonyms, additionalkeywords),
+                                     query = create_query_string_T_scopus(genus,
+                                                                          species,
+                                                                          synonyms,
+                                                                          additionalkeywords),
                                      httpAccept = "application/xml")) #format the URL to be sent to the API
   httr::stop_for_status(response) #pass any HTTP errors to the R console
   response_data <- XML::xmlParse(response) #parse the data to extract values
-  resultCount <- as.numeric(XML::xpathSApply(response_data,"//opensearch:totalResults", XML::xmlValue)) #get the total number of search results for the string
+  resultCount <- as.numeric(XML::xpathSApply(response_data,
+                                             "//opensearch:totalResults",
+                                             XML::xmlValue)) #get the total number of search results for the string
   return(resultCount)
 }
 
@@ -220,11 +225,16 @@ CountSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, data
   sp_check(genus, species)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
-                                     query = create_query_string_TAK_scopus(genus, species, synonyms, additionalkeywords),
+                                     query = create_query_string_TAK_scopus(genus,
+                                                                            species,
+                                                                            synonyms,
+                                                                            additionalkeywords),
                                      httpAccept = "application/xml")) #format the URL to be sent to the API
   httr::stop_for_status(response) #pass any HTTP errors to the R console
   response_data <- XML::xmlParse(response) #parse the data to extract values
-  resultCount <- as.numeric(XML::xpathSApply(response_data,"//opensearch:totalResults", XML::xmlValue)) #get the total number of search results for the string
+  resultCount <- as.numeric(XML::xpathSApply(response_data,
+                                             "//opensearch:totalResults",
+                                             XML::xmlValue)) #get the total number of search results for the string
   return(resultCount)
 }
 
@@ -269,11 +279,15 @@ CountGenusT_scopus <- function(genus, synonyms, additionalkeywords, datatype = "
   genus_check(genus)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
-                                     query = create_query_string_T_scopus_genus(genus, synonyms, additionalkeywords),
+                                     query = create_query_string_T_scopus_genus(genus,
+                                                                                synonyms,
+                                                                                additionalkeywords),
                                      httpAccept = "application/xml")) #format the URL to be sent to the API
   httr::stop_for_status(response) #pass any HTTP errors to the R console
   response_data <- XML::xmlParse(response) #parse the data to extract values
-  resultCount <- as.numeric(XML::xpathSApply(response_data,"//opensearch:totalResults", XML::xmlValue)) #get the total number of search results for the string
+  resultCount <- as.numeric(XML::xpathSApply(response_data,
+                                             "//opensearch:totalResults",
+                                             XML::xmlValue)) #get the total number of search results for the string
   return(resultCount)
 }
 
@@ -318,11 +332,15 @@ CountGenusTAK_scopus <- function(genus, synonyms, additionalkeywords, datatype =
   genus_check(genus)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
-                                     query = create_query_string_TAK_scopus_genus(genus, synonyms, additionalkeywords),
+                                     query = create_query_string_TAK_scopus_genus(genus,
[…]
