### 2021-12-13 #3776 @jessicatytam
Thanks everyone for your helpful comments! I've since updated the documentation, reorganised the functions, and fixed some of the warnings that were showing up in checks. We're currently working on adding more tests for the API.

In regards to @kamapu 's comments:
- I've replaced all the `Fetch*()` and `Count*()` functions with `Fetch()` and `Count()` to make things simpler
- the error `Bad Request (HTTP 400)` is probably a result of your institution not being a subscriber of Elsevier/Scopus (see https://github.com/jessicatytam/specieshindex/issues/46#issue-1029956107)
- when you say `obj <- Fetch()` and `obj <- Count()` are not working, do you mean they are returning some sort of error? They are both working for me, do you want to provide a reprex?
