### 2021-12-15 5d1ba4c
adding tests for wos requests
