+++ R/specieshindex.R
@@ -166,7 +166,11 @@ Fetch <- function(db,
 #' 
 #' @noRd
 #' 
-CountSpT_scopus <- function(genus, species, synonyms, additionalkeywords, datatype = "application/xml") {
+CountSpT_scopus <- function(genus,
+                            species,
+                            synonyms,
+                            additionalkeywords,
+                            datatype = "application/xml") {
   sp_check(genus, species)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
@@ -221,7 +225,11 @@ CountSpT_scopus <- function(genus, species, synonyms, additionalkeywords, dataty
 #' 
 #' @noRd
 #' 
-CountSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, datatype = "application/xml") {
+CountSpTAK_scopus <- function(genus,
+                              species,
+                              synonyms,
+                              additionalkeywords,
+                              datatype = "application/xml") {
   sp_check(genus, species)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
@@ -275,7 +283,10 @@ CountSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, data
 #' 
 #' @noRd
 #' 
-CountGenusT_scopus <- function(genus, synonyms, additionalkeywords, datatype = "application/xml") {
+CountGenusT_scopus <- function(genus,
+                               synonyms,
+                               additionalkeywords,
+                               datatype = "application/xml") {
   genus_check(genus)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
@@ -328,7 +339,10 @@ CountGenusT_scopus <- function(genus, synonyms, additionalkeywords, datatype = "
 #' 
 #' @noRd
 #' 
-CountGenusTAK_scopus <- function(genus, synonyms, additionalkeywords, datatype = "application/xml") {
+CountGenusTAK_scopus <- function(genus,
+                                 synonyms,
+                                 additionalkeywords,
+                                 datatype = "application/xml") {
   genus_check(genus)
   response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                         query = list(apiKey = apikey,
@@ -379,8 +393,15 @@ CountGenusTAK_scopus <- function(genus, synonyms, additionalkeywords, datatype =
 #' 
 #' @noRd
 #' 
-FetchSpT_scopus <- function(genus, species, synonyms, additionalkeywords, language = 0) {
-  count <- CountSpT_scopus(genus, species, synonyms, additionalkeywords) #check the number of records
+FetchSpT_scopus <- function(genus,
+                            species,
+                            synonyms,
+                            additionalkeywords,
+                            language = 0) {
+  count <- CountSpT_scopus(genus,
+                           species,
+                           synonyms,
+                           additionalkeywords) #check the number of records
   print(paste(count, "records found."))
   if (count < 1) {
     noCitations <- data.frame(citations = 0)
@@ -627,8 +648,15 @@ FetchSpT_scopus <- function(genus, species, synonyms, additionalkeywords, langua
 #' 
 #' @noRd
 #' 
-FetchSpTAK_scopus <- function(genus, species, synonyms, additionalkeywords, language = 0) {
-  count <- CountSpTAK_scopus(genus, species, synonyms, additionalkeywords) #check the number of records
+FetchSpTAK_scopus <- function(genus,
+                              species,
+                              synonyms,
+                              additionalkeywords,
+                              language = 0) {
+  count <- CountSpTAK_scopus(genus,
+                             species,
+                             synonyms,
+                             additionalkeywords) #check the number of records
   print(paste(count, "records found."))
   if (count < 1) {
[…]
