+++ man/SpMindex.Rd
@@ -17,7 +17,12 @@ This function calculates the m-index of  species.
 M-index uses the h-index and divides it by the number of years of activity.
 }
 \examples{
-data(Woylie)
+\dontrun{
+Woylie <- Fetch(db = "scopus",
+                search = "tak",
+                level = "species",
+                genus = "Bettongia", species = "penicillata")
+}
 SpMindex(Woylie)
 
 }
