### 2021-12-15 0b8a3d0
adding tests for BASE
