### 2022-02-08 ac97ab1
plotting m-index
