+++ R/model_without_domesticated_spp.R
@@ -1,0 +1,7 @@
+#load data
+includeh <- read.csv(file = "outputs/data/includeh.csv")[-c(1)]
+
+includeh[includeh$domestication_bin==1,] #domesticated spp
+sum(includeh$domestication_bin==1) #12
+includeh[includeh$domestication_bin==2,] #partially-domesticated spp
+sum(includeh$domestication_bin==2) #136
