### 2022-02-22 #3776 @jessicatytam
Hi @Bisaloo apologies for the holdup! I've been working on multiple projects and getting organised as I'm starting my PhD. We hope to finishing addressing the comments by the end of March. Here's a list of tasks we will be working on:

- [ ] Adding API tests
- [ ] Reduce code duplication
- [ ] Use `facet_wrap` in plots
- [ ] Add more error messages
- [ ] Implement a secure way to provide credentials in sessions
