### 2022-02-23 4a018c1
trying to replace ggarrange with facet_wrap, having some difficulty with varying decimal places in the same column
