+++ tests/testthat/test_scopus.R
@@ -9,12 +9,14 @@ with_mock_api({ #pass
   })
 })
 
-# test_that("CountSpT_scopus_works", {
-#   response_data <- XML::xmlParse(t_scopus_xml) #parse the data to extract values
-#   resultCount <- as.numeric(XML::xpathSApply(t_scopus_xml,"//opensearch:totalResults", XML::xmlValue))
-#   expect_equal(CountSpT("scopus", "Bettongia", "penicillata"), resultCount)
-#   expect_identical(CountSpT("scopus", "Bettongia", "penicillata"), CountSpT_scopus("Bettongia", "penicillata"))
-# })
+test_that("CountSpT_scopus_works", { #Error: XML content does not seem to be XML: 'ot exist."}}'
+  response_data <- XML::xmlParse(t_scopus) #parse the data to extract values
+  resultCount <- as.numeric(XML::xpathSApply(t_scopus,"//opensearch:totalResults", XML::xmlValue))
+  expect_equal(CountSpT("scopus", "Bettongia", "penicillata"), resultCount)
+  expect_identical(CountSpT("scopus", "Bettongia", "penicillata"), CountSpT_scopus("Bettongia", "penicillata"))
+})
+
+fromJSON((rawToChar(t_scopus$content)))
 
 with_mock_api({
   test_that("FetchSpT_scopus_works", {
