+++ specieshindex.Rmd
@@ -769,22 +769,27 @@ genus_check <- function(genus) {
 ```{r}
 #theme for ggplot
 spindex_plot_theme <- function() {
-  ggplot2::theme(title = element_text(size = 12,
-                                      face = "bold"),
+  ggplot2::theme(title = element_blank(),
                  axis.title.x = element_blank(),
                  axis.text.x = element_blank(),
                  axis.ticks.x = element_blank(),
                  axis.line.x = element_line(colour = "grey20"),
                  axis.title.y = element_blank(),
                  axis.text.y = element_text(size = 10),
+                 strip.background = element_rect(fill = "white"),
+                 strip.text = element_text(size = 16,
+                                           face = "bold",
+                                           hjust = -0.02),
+                 legend.title = element_text(size = 12,
+                                             face = "bold"),
                  legend.key = element_rect(fill = "white"),
                  plot.background = element_rect(fill = "white"),
                  panel.background = element_rect(fill = "white"),
                  panel.grid.major.y = element_line(colour = "grey90"),
                  panel.grid.minor.y = element_line(colour = "grey90",
                                                    linetype = "longdash"),
-                 panel.grid.major.x = element_blank(),
-                 panel.grid.minor.x = element_blank())
+                 panel.grid.major.x = ggplot2::element_blank(),
+                 panel.grid.minor.x = ggplot2::element_blank())
 }
 ```
 
@@ -808,22 +813,6 @@ sppub_plot_theme <- function() {
                                                    linetype = "longdash"),
                  panel.grid.major.x = ggplot2::element_blank(),
                  panel.grid.minor.x = ggplot2::element_blank())
-}
-```
-
-```{r}
-individualIndex <- function(data, y, title) {
-  ifelse(w$index=="m-index", round(w$value, 3), round(w$value, 0))
-  p <- ggplot2::ggplot(data = data,
-                       ggplot2::aes(x = genus_species,
-                                    {{y}},
-                                    colour = genus_species)) +
-  ggplot2::geom_point(size = 2) +
-  ggplot2::labs(title = paste0(title),
-                colour = "Species") ++
-  scale_y_continuous(labels = scales::label_number(drop0trailing = T)) +
-    spindex_plot_theme()
-  return(p)
 }
 ```
 
@@ -1984,20 +1973,25 @@ Allindices <- function(data, genus, species, sourcetype = 0) {
 ```{r}
 #plotting data from Allindices()
 plotAllindices <- function(data) {
-  ifelse(data$index=="m-index", round(data$value, 3), round(data$value, 0))
-  #h-index
-  h_plot <- individualIndex(data, h, "h-index")
-  #m-index
-  m_plot <- individualIndex(data, m, "m-index")
-  #i10
-  i10_plot <- individualIndex(data, i10, "i10 index")
-  #h5
-  h5_plot <- individualIndex(data, h5, "h5 index")
-  #combining the plots
-  combine_plot <- ggpubr::ggarrange(h_plot, m_plot, i10_plot, h5_plot,
-                                    common.legend = TRUE,
-                                    legend = "right")
-  return(combine_plot)
+  facet_data <- data %>% 
+    tidyr::pivot_longer(cols = h:h5,
+                        names_to = "index",
+                        values_to = "value")
+  facet_data$index <- factor(facet_data$index,
+                             levels = c("h", "m", "i10", "h5"),
+                             labels = c("h-index", "m-index", "i10 index", "h5 index"))
+  p <- ggplot(data = facet_data,
[…]
