### 2022-02-25 a3fcbcd
new plots without domesticated spp
