### 2022-02-25 41cd06e
fixed y-axis of new plots
