### 2022-02-25 6de27b5
new grid plot showing domestication status
