### 2022-02-25 b394fec
new colour palette for new plot
