### 2022-02-28 dcd5a9b
condensed scopus cCount functions
