+++ R/specieshindex.R
@@ -134,8 +134,9 @@ Fetch <- function(db,
 #' It counts the publications with the binomial name in the title only.
 #' A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
 #' 
-#' @title Search count from Scopus - title only
+#' @title Search count from Scopus 
 #'
+#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
 #' @param synonyms Alternate species names.
@@ -144,217 +145,54 @@ Fetch <- function(db,
 #'
 #' @return Search count of the species with the given \code{genus} and \code{species}.
 #' 
-#' 
 #' @references 
 #' Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
-#'
+#' 
 #' @examples
 #' \dontrun{
-#' CountSpT_scopus("Bettongia", "penicillata")
+#' Count_scopus(search = "t", genus = "Bettongia", species = "penicillata")
 #' 
 #' #lower case letter in genus is also accepted and will return identical results
 #' 
-#' CountSpT_scopus("bettongia", "penicillata")
+#' Count_scopus(search = "t", genus = "Bettongia", species = "penicillata")
 #' }
 #' \dontrun{
-#' CountSpT_scopus("Bettongia", "penicillata", "conserv*")
+#' Count_scopus(search = "t", genus = "Bettongia", species = "penicillata", additionalkeywords = "conserv*")
 #' 
 #' #lower case letter in genus is also accepted and will return identical results
 #' 
-#' CountSpT_scopus("bettongia", "penicillata", "conserv*")
+#' Count_scopus(search = "t", genus = "bettongia", species = "penicillata", additionalkeywords = "conserv*")
 #' }
 #' 
 #' @noRd
 #' 
-CountSpT_scopus <- function(genus,
-                            species,
-                            synonyms,
-                            additionalkeywords,
-                            datatype = "application/xml") {
-  sp_check(genus, species)
-  response <- httr::GET("http://api.elsevier.com/content/search/scopus",
-                        query = list(apiKey = apikey,
-                                     query = create_query_string_T_scopus(genus,
-                                                                          species,
-                                                                          synonyms,
-                                                                          additionalkeywords),
-                                     httpAccept = "application/xml")) #format the URL to be sent to the API
-  httr::stop_for_status(response) #pass any HTTP errors to the R console
-  response_data <- XML::xmlParse(response) #parse the data to extract values
-  resultCount <- as.numeric(XML::xpathSApply(response_data,
-                                             "//opensearch:totalResults",
-                                             XML::xmlValue)) #get the total number of search results for the string
-  return(resultCount)
-}
-
-
-
-#' This function counts the total number of search results.
-#' It counts the publications with the binomial name in the title, abstract, and keywords.
-#' A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
-#' 
-#' @title Search count from Scopus - title, abstract, and keywords
-#'
-#' @param genus Genus classification from the binomial name.
-#' @param species Species classification from the binomial name.
-#' @param synonyms Alternate species names.
-#' @param additionalkeywords Optional search terms.
-#' @param datatype Formats the URL to be sent to the API. The default is "application/xml".
-#'
-#' @return Search count of the species with the given \code{genus} and \code{species}.
-#' 
[…]
