+++ specieshindex.Rmd
@@ -731,6 +731,32 @@ create_query_string_TAK_lens_genus <- function(genus, synonyms, additionalkeywor
 ```
 
 ```{r}
+scopus_request_t <- function(request) {
+  rscopus::scopus_search(query = paste0(create_query_string_T_scopus(genus,
+                                                                     species,
+                                                                     synonyms,
+                                                                     additionalkeywords),
+                                        paste0(request)),
+                         api_key = apikey,
+                         verbose = TRUE,
+                         wait_time = 3)
+}
+```
+
+```{r}
+scopus_request_tak <- function(request) {
+  rscopus::scopus_search(query = paste0(create_query_string_TAK_scopus(genus,
+                                                                       species,
+                                                                       synonyms,
+                                                                       additionalkeywords),
+                                        paste0(request)),
+                         api_key = apikey,
+                         verbose = TRUE,
+                         wait_time = 3)
+}
+```
+
+```{r}
 #species check
 sp_check <- function(genus, species) {
   findname <- taxize::gnr_resolve(sci = paste(genus, species),
@@ -762,6 +788,24 @@ genus_check <- function(genus) {
     print(paste("Genus found on CoL, ITIS, NCBI, or EoL."))
   } else {
     stop("Genus not found on CoL, ITIS, NCBI, or EoL. Please check your spelling and try again.")
+  }
+}
+```
+
+```{r}
+#species check
+test_check <- function(genus, species = NULL) {
+  findname <- taxize::gnr_resolve(sci = paste(genus, species = paste(species)),
+                                  data_source_ids = list("1", "3", "4", "12")) #check if the species exist
+  if (length(findname>0)) {
+    findname <- findname[order(-findname$score),]
+  } else {
+    stop("Species not found on CoL, ITIS, NCBI, or EoL. Please check your spelling and try again.")
+  }
+  if (findname$score[1]>=0.75) {
+    print(paste("Species found on CoL, ITIS, NCBI, or EoL."))
+  } else {
+    stop("Species not found on CoL, ITIS, NCBI, or EoL. Please check your spelling and try again.")
   }
 }
 ```
@@ -874,6 +918,38 @@ Fetch <- function(db, search, level, genus, species, synonyms, additionalkeyword
 }
 ```
 
+```{r, this works}
+Count_scopus <- function(level,
+                         genus,
+                         species = NULL,
+                         synonyms,
+                         additionalkeywords,
+                         datatype = "application/xml") {
+  test_check(genus, species = paste(species))
+  if (search == "t") {
+    response <- httr::GET("http://api.elsevier.com/content/search/scopus",
+                      query = list(apiKey = apikey,
+                                   query = create_query_string_T_scopus(genus,
+                                                                        species = paste(species),
+                                                                        synonyms,
+                                                                        additionalkeywords),
+                                   httpAccept = "application/xml")) #format the URL to be sent to the API
+  } else if (search == "tak") {
+    response <- httr::GET("http://api.elsevier.com/content/search/scopus",
+                      query = list(apiKey = apikey,
+                                   query = create_query_string_TAK_scopus(genus,
+                                                                          species = paste(species),
+                                                                          synonyms,
+                                                                          additionalkeywords),
+                                   httpAccept = "application/xml")) #format the URL to be sent to the API
+  }
+  httr::stop_for_status(response) #pass any HTTP errors to the R console
[…]
