### 2022-02-28 de72d42
condensed scopus Fetch functions and search strings for all db
