+++ R/specieshindex.R
@@ -172,12 +172,13 @@ Count_scopus <- function(search,
                          synonyms,
                          additionalkeywords,
                          datatype = "application/xml") {
-  test_check(genus, species = paste(species))
+  sp_check(genus,
+           species = paste0(species))
   if (search == "t") {
     response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                           query = list(apiKey = apikey,
                                        query = create_query_string_T_scopus(genus,
-                                                                            species = paste(species),
+                                                                            species = paste0(species),
                                                                             synonyms,
                                                                             additionalkeywords),
                                        httpAccept = "application/xml")) #format the URL to be sent to the API
@@ -185,10 +186,12 @@ Count_scopus <- function(search,
     response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                           query = list(apiKey = apikey,
                                        query = create_query_string_TAK_scopus(genus,
-                                                                              species = paste(species),
+                                                                              species = paste0(species),
                                                                               synonyms,
                                                                               additionalkeywords),
                                        httpAccept = "application/xml")) #format the URL to be sent to the API
+  } else {
+    stop('Set search = "t" for title-only searches, or "tak" for searches in the title, abstract, and keywords.')
   }
   httr::stop_for_status(response) #pass any HTTP errors to the R console
   response_data <- XML::xmlParse(response) #parse the data to extract values
@@ -231,13 +234,14 @@ Count_scopus <- function(search,
 #' 
 #' @noRd
 #' 
-FetchSpT_scopus <- function(genus,
-                            species,
-                            synonyms,
-                            additionalkeywords,
-                            language = 0) {
-  count <- CountSpT_scopus(genus,
-                           species,
+FetchT_scopus <- function(genus,
+                          species = NULL,
+                          synonyms,
+                          additionalkeywords,
+                          language = 0) {
+  count <- CountSpT_scopus(search = "t",
+                           genus,
+                           species = paste0(species),
                            synonyms,
                            additionalkeywords) #check the number of records
   print(paste(count, "records found."))
@@ -252,7 +256,7 @@ FetchSpT_scopus <- function(genus,
       response <- httr::GET("http://api.elsevier.com/content/search/scopus",
                             query = list(apiKey = apikey,
                                          query = paste0(create_query_string_T_scopus(genus,
-                                                                                     species,
+                                                                                     species = paste0(species),
                                                                                      synonyms,
                                                                                      additionalkeywords),
                                                         " AND LANGUAGE(", lang$language[j], ")"),
@@ -268,7 +272,8 @@ FetchSpT_scopus <- function(genus,
         for (i in 2019:1990) {
           searchloop <- scopus_request_t('" AND PUBYEAR = ", i, " AND LANGUAGE(", lang$language[j], ")"')
           searchlooplist <- rscopus::entries_to_citation_df(searchloop$entries)
-          searchloopdf <- dplyr::bind_rows(searchloopdf, searchlooplist)
+          searchloopdf <- dplyr::bind_rows(searchloopdf,
+                                           searchlooplist)
         }
         search1985 <- scopus_request_t('" AND PUBYEAR > 1984 AND PUBYEAR < 1990 AND LANGUAGE(", lang$language[j], ")"')
         search1985df <- rscopus::entries_to_citation_df(search1985$entries)
@@ -289,7 +294,8 @@ FetchSpT_scopus <- function(genus,
                                      search_olddf)
[…]
