+++ specieshindex.Rmd
@@ -25,15 +25,18 @@ library(specieshindex)
 ```
 
 ```{r, scopus query for title only searches}
-create_query_string_T_scopus <- function(genus, species, synonyms, additionalkeywords){
+create_query_string_T_scopus <- function(genus,
+                                         species = NULL,
+                                         synonyms,
+                                         additionalkeywords){
   if (missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE("', genus, ' ', species, '")'))
+    return(paste0('TITLE("', genus, ' ', species = paste0(species), '")'))
   } 
   if (!missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE("', genus, ' ', species, '" AND ', additionalkeywords, ')'))
+    return(paste0('TITLE("', genus, ' ', species = paste0(species), '" AND ', additionalkeywords, ')'))
   }
   if (missing(additionalkeywords) & !missing(synonyms)) {
-    temp_string <- paste0('TITLE("', genus, ' ', species, '" OR "', synonyms[1], '"')
+    temp_string <- paste0('TITLE("', genus, ' ', species = paste0(species), '" OR "', synonyms[1], '"')
     if (length(synonyms)==1) {
       return(paste0(temp_string, ')'))
     }
@@ -45,21 +48,25 @@ create_query_string_T_scopus <- function(genus, species, synonyms, additionalkey
     }
   }
   if (!missing(additionalkeywords) & !missing(synonyms)) {
-    return(paste0('TITLE(("', genus, ' ', species, '" OR "', synonyms, '") AND ', additionalkeywords, ')'))
+    return(paste0('TITLE(("', genus, ' ', species = paste0(species), '" OR "',
+                  synonyms, '") AND ', additionalkeywords, ')'))
   } 
 }
 ```
 
 ```{r, scopus query for title, abstract, and keywords searches}
-create_query_string_TAK_scopus <- function(genus, species, synonyms, additionalkeywords){
+create_query_string_TAK_scopus <- function(genus,
+                                           species = NULL,
+                                           synonyms,
+                                           additionalkeywords){
   if (missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE-ABS-KEY("', genus, ' ', species, '")'))
+    return(paste0('TITLE-ABS-KEY("', genus, ' ', species = paste0(species), '")'))
   } 
   if (!missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE-ABS-KEY("', genus, ' ', species, '" AND ', additionalkeywords, ')'))
+    return(paste0('TITLE-ABS-KEY("', genus, ' ', species = paste0(species), '" AND ', additionalkeywords, ')'))
   }
   if (missing(additionalkeywords) & !missing(synonyms)) {
-    temp_string <- paste0('TITLE-ABS-KEY("', genus, ' ', species, '" OR "', synonyms[1], '"')
+    temp_string <- paste0('TITLE-ABS-KEY("', genus, ' ', species = paste0(species), '" OR "', synonyms[1], '"')
     if (length(synonyms)==1) {
       return(paste0(temp_string, ')'))
     }
@@ -71,7 +78,8 @@ create_query_string_TAK_scopus <- function(genus, species, synonyms, additionalk
     }
   }
   if (!missing(additionalkeywords) & !missing(synonyms)) {
-    return(paste0('TITLE-ABS-KEY(("', genus, ' ', species, '" OR "', synonyms, '") AND ', additionalkeywords, ')'))
+    return(paste0('TITLE-ABS-KEY(("', genus, ' ', species = paste0(species), '" OR "',
[…]
