### 2022-03-01 64c2dfa
condensed all API functions; wos needs more testing, updated manual
