+++ R/specieshindex.R
@@ -5,7 +5,6 @@
 #'
 #' @param db Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").
 #' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
-#' @param level Taxonomic level. Genus ("genus") or species ("species").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
 #' @param synonyms Alternate species names.
@@ -18,13 +17,11 @@
 #' \dontrun{
 #' Count(db = "scopus",
 #'       search = "t",
-#'       level = "species",
 #'       genus = "Osphranter", species = "rufus")
 #' }
 #' \dontrun{
 #' Count(db = "scopus",
 #'       search = "t",
-#'       level = "species",
 #'       genus = "Osphranter", species = "rufus",
 #'       synonyms = "Macropus rufus",
 #'       additionalkeywords = "conserv*")
@@ -33,35 +30,23 @@ Count <- function(db,
                   search,
                   level,
                   genus,
-                  species,
+                  species = NULL,
                   synonyms,
                   additionalkeywords) {
-if (db == "scopus" & search == "t" & level == "species") {
-  countsp <- CountSpT_scopus(genus, species, synonyms, additionalkeywords)
-} else if (db == "wos" & search == "t" & level == "species") {
-  countsp <- CountSpT_wos(genus, species, synonyms, additionalkeywords)
-} else if (db == "base" & search == "t" & level == "species") {
-  countsp <- CountSpT_base(genus, species, synonyms, additionalkeywords)
-} else if (db == "scopus" & search == "tak" & level == "species") {
-  countsp <- CountSpTAK_scopus(genus, species, synonyms, additionalkeywords)
-} else if (db == "wos" & search == "tak" & level == "species") {
-  countsp <- CountSpTAK_wos(genus, species, synonyms, additionalkeywords)
-} else if (db == "base" & search == "tak" & level == "species") {
-  countsp <- CountSpTAK_base(genus, species, synonyms, additionalkeywords)
-} else if (db == "scopus" & search == "t" & level == "genus") {
-  countsp <- CountGenusT_scopus(genus, synonyms, additionalkeywords)
-} else if (db == "wos" & search == "t" & level == "genus") {
-  countsp <- CountGenusT_wos(genus, synonyms, additionalkeywords)
-} else if (db == "base" & search == "t" & level == "genus") {
-  countsp <- CountGenusT_base(genus, synonyms, additionalkeywords)
-} else if (db == "scopus" & search == "tak" & level == "genus") {
-  countsp <- CountGenusTAK_scopus(genus, synonyms, additionalkeywords)
-} else if (db == "wos" & search == "tak" & level == "genus") {
-  countsp <- CountGenusTAK_wos(genus, synonyms, additionalkeywords)
-} else if (db == "base" & search == "tak" & level == "genus") {
-  countsp <- CountGenusTAK_base(genus, synonyms, additionalkeywords)
-} 
-return(countsp)
+  if (db == "scopus" & search == "t") {
+    countsp <- Count_scopus(search = "t", genus, species, synonyms, additionalkeywords)
+  } else if (db == "scopus" & search == "tak") {
+    countsp <- Count_scopus(search = "tak", genus, species, synonyms, additionalkeywords)
+  } else if (db == "wos" & search == "t") {
+    countsp <- Count_wos(saerch = "t", genus, species, synonyms, additionalkeywords)
+  } else if (db == "wos" & search == "tak") {
+    countsp <- Count_wos(saerch = "tak", genus, species, synonyms, additionalkeywords)
+  } else if (db == "base" & search == "t") {
[…]
