+++ README.rmd
@@ -104,13 +104,11 @@ Multiple databases have been incorporated into `specieshindex`, namely Scopus, W
 #Title only; species level
 Count(db = "scopus",
       search = "t",
-      level = "species",
       genus = "Bettongia", species = "penicillata")
 
 #Title, abstract, and keywords; genus level
 Count(db = "scopus",
       search = "tak",
-      level = "genus",
       genus = "Bettongia")
 ```
 
@@ -121,19 +119,15 @@ In order to calculate the indices, you will need to download the citation record
 ```{r, eval=FALSE}
 Woylie <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Bettongia", species = "penicillata")
 Quokka <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Setonix", species = "brachyurus")
 Platypus <- Fetch(db = "scopus",
                   search = "tak",
-                  level = "species",
                   genus = "Ornithorhynchus", species = "anatinus")
 Koala <- Fetch(db = "scopus",
                search = "tak",
-               level = "species",
                genus = "Phascolarctos", species = "cinereus")
 ```
 
