### 2022-03-01 56d0dd5
updated documentation and added instructions to securely use their api key
