+++ vignettes/specieshindex.Rmd
@@ -94,11 +94,29 @@ Before initiating the retrieval of any data, make sure you have registered an AP
 3. Go to the `My API Key` tab on top of the page and click `Create API Key`.
 4. Read the legal documents and check the boxes.
 
-Run the following line of code to set your API key for the current session and get started:
+After acquiring your key, make sure to store it safely. The following steps will enable you to save it as an environment variable.
 
 ```{r, eval=FALSE}
-#API key
-apikey <- "your_api_key_from_scopus"
+file.edit("~/.Renviron")
+```
+
+This will bring up an empty file, which is where you will save your key into.
+
+```{r, eval=FALSE}
+scopus_api_key = "a_long_string"
+```
+
+Restart your session for this to work. To retrieve your key, use `Sys.getenv()`:
+
+```{r, eval=FALSE}
+Sys.getenv("scopus_api_key")
+#> [1] "a_long_string"
+```
+
+You can then load it to your environment without having the secret key in your script or console as follows:
+
+```{r}
+apikey <- Sys.getenv("scopus_api_key")
 ```
 
 ## Connecting to Web of Science
@@ -121,7 +139,7 @@ Bielefeld Academic Search Engine (BASE) is a literature database that contains b
 
 The searches are performed using species Latin name. The relevant functions take either a Latin binominal species name as their main parameters, e.g. "*Bettongia*", "*penicillata*”, or their genus name, e.g. "*Bettongia*". Make sure you are using binomial names instead of common names as required by the code.
 
-We have implemented functions that check the available number of relevant records (`Count()`) and functions that download these records (`Fetch()`) for a given species. `specieshindex` supports multiple databases, including Scopus, Web of Science, and BASE. To differentiate between them, you can change the `db` parameter. You can set `search = "t"` for search terms in the title only and `search = "tak"` for search terms in the title, abstract, or keywords. To find articles of species, set `level = "species"`. To find articles of genera, set `level = "genus"`.
+We have implemented functions that check the available number of relevant records (`Count()`) and functions that download these records (`Fetch()`) for a given species. `specieshindex` supports multiple databases, including Scopus, Web of Science, and BASE. To differentiate between them, you can change the `db` parameter. You can set `search = "t"` for search terms in the title only and `search = "tak"` for search terms in the title, abstract, or keywords. For genus-level searches, leave the `species` parameter empty. 
 
 We recommend that you run a quick search using the `Count()` function before running `Fetch()` to extract any data since this can be a tedious process for searches with many results. Both functions give a search count of the number of records found on the databases.
 
@@ -244,7 +262,6 @@ For example, if you want to retrieve documents with the language variable, you c
 ```{r, eval=FALSE}
 Fetch(db = "scopus",
       search = "t",
-      level = "species",
       genus = "Ailuropoda", species = "melanoleuca",
       language = 1)
 ```
