### 2022-03-01 1578ea0
trying to get windows to behave in github actions
