+++ .github/workflows/main.yml
@@ -33,7 +33,7 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v2.3.4
 
       - uses: r-lib/actions/setup-r@v1
         with:
@@ -91,7 +91,7 @@ jobs:
          GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
        steps:
-         - uses: actions/checkout@v2
+         - uses: actions/checkout@v2.3.4
 
          - uses: r-lib/actions/setup-r@v1
            with:
