+++ .github/workflows/main.yml
@@ -33,7 +33,7 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
-      - uses: actions/checkout@v2.3.4
+      - uses: actions/checkout@v2
 
       - uses: r-lib/actions/setup-r@v1
         with:
@@ -52,7 +52,7 @@ jobs:
 
       - name: Cache R packages
         if: runner.os != 'Windows'
-        uses: actions/checkout@v2.3.4
+        uses: actions/checkout@v2
         with:
           path: ${{ env.R_LIBS_USER }}
           key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
@@ -91,7 +91,7 @@ jobs:
          GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
        steps:
-         - uses: actions/checkout@v2.3.4
+         - uses: actions/checkout@v2
 
          - uses: r-lib/actions/setup-r@v1
            with:
