+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: 2.7.3
+pkgdown: 2.0.2
+pkgdown_sha: ~
+articles:
+  specieshindex: specieshindex.html
+last_built: 2022-03-01T04:13Z
+
