+++ man/Platypus.Rd
@@ -1,0 +1,42 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/specieshindex.R
+\docType{data}
+\name{Platypus}
+\alias{Platypus}
+\title{Platypus}
+\format{
+A data frame :
+\describe{
+  \item{citations}{number of cites}
+  \item{journal}{journal}
+  \item{authkeywords}{authkeywords}
+  \item{cover_date}{cover_date}
+  \item{cover_display_date}{cover_display_date}
+  \item{dc_creator}{dc_creator}
+  \item{dc_description}{dc_description}
+  \item{dc_identifier}{dc_identifier}
+  \item{description}{description}
+  \item{doi}{doi}
+  \item{eid}{eid}
+  \item{pii}{pii}
+  \item{prism_aggregationType}{prism_aggregationType}
+  \item{prism_eIssn}{prism_eIssn}
+  \item{prism_issn}{prism_issn}
+  \item{prism_pageRange}{prism_pageRange}
+  \item{prism_url}{prism_url}
+  \item{source_id}{source_id}
+  \item{subtype}{subtype}
+  \item{title}{title}
+  ...
+}
+}
+\source{
+WOS
+}
+\usage{
+Platypus
+}
+\description{
+A dataset with some literature on Platypus
+}
+\keyword{datasets}

[…]
