### 2022-03-01 d8b2dbb
updated README and vignette
