+++ README.rmd
@@ -49,7 +49,9 @@ To connect and download citation information from Scopus legally, you will **abs
 3. Go to the `My API Key` tab on top of the page and click `Create API Key`.
 4. Read the legal documents and check the boxes. 
 
-After acquiring your key, make sure to store it safely. The following steps will enable you to save it as an environment variable.
+#### Using your API key securely
+
+After acquiring your key, make sure to store it safely. The following steps will enable you to save it as an environment variable, without saving it in the console or script.
 
 ```{r, eval=FALSE}
 file.edit("~/.Renviron")
@@ -68,7 +70,7 @@ Sys.getenv("scopus_api_key")
 #> [1] "a_long_string"
 ```
 
-You can then load it to your environment without having the secret key in your script or console as follows:
+You can then load it to your environment as follows:
 
 ```{r}
 apikey <- Sys.getenv("scopus_api_key")
