### 2022-03-01 33c0e50
did some checking and fixed specieshindex.R
