+++ R/NULLing.R
@@ -11,3 +11,7 @@ i10 <- NULL
 h5 <- NULL
 Freq <- NULL
 spp <- NULL
+genus <- NULL
+synonyms <- NULL
+additionalkeywords <- NULL
+value <- NULL
