+++ specieshindex.Rmd
@@ -84,58 +84,6 @@ create_query_string_TAK_scopus <- function(genus,
 }
 ```
 
-```{r, scopus query for title only searches; genus}
-create_query_string_T_scopus_genus <- function(genus, synonyms, additionalkeywords){
-  if (missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE("', genus, '")'))
-  } 
-  if (!missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE("', genus, '" AND ', additionalkeywords, ')'))
-  }
-  if (missing(additionalkeywords) & !missing(synonyms)) {
-    temp_string <- paste0('TITLE("', genus, '" OR "', synonyms[1], '"')
-    if (length(synonyms)==1) {
-      return(paste0(temp_string, ')'))
-    }
-    else {
-      for (i in 2:length(synonyms)){
-        temp_string <- paste0(temp_string, ' OR "', synonyms[i], '"')
-      }
-      return(paste0(temp_string, ')'))
-    }
-  }
-  if (!missing(additionalkeywords) & !missing(synonyms)) {
-    return(paste0('TITLE(("', genus, '" OR "', synonyms, '") AND ', additionalkeywords, ')'))
-  } 
-}
-```
-
-```{r, scopus query for title, abstract, and keywords searches; genus}
-create_query_string_TAK_scopus_genus <- function(genus, synonyms, additionalkeywords){
-  if (missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE-ABS-KEY("', genus, '")'))
-  } 
-  if (!missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TITLE-ABS-KEY("', genus, '" AND ', additionalkeywords, ')'))
-  }
-  if (missing(additionalkeywords) & !missing(synonyms)) {
-    temp_string <- paste0('TITLE-ABS-KEY("', genus, '" OR "', synonyms[1], '"')
-    if (length(synonyms)==1) {
-      return(paste0(temp_string, ')'))
-    }
-    else {
-      for (i in 2:length(synonyms)){
-         temp_string <- paste0(temp_string, ' OR "', synonyms[i], '"')
-      }
-      return(paste0(temp_string, ')'))
-    }
-  }
-  if (!missing(additionalkeywords) & !missing(synonyms)) {
-    return(paste0('TITLE-ABS-KEY(("', genus, '" OR "', synonyms, '") AND ', additionalkeywords, ')'))
-  } 
-}
-```
-
 ```{r, wos query for title only searches}
 create_query_string_T_wos <- function(genus,
                                       species = NULL,
@@ -204,68 +152,6 @@ create_query_string_TAK_wos <- function(genus,
                   synonyms, '") AND ', additionalkeywords, ')',
                   ' OR AK = (("', genus, ' ', species = paste0(species), '" OR "',
                   synonyms, '") AND ', additionalkeywords, ')'))
-  } 
-}
-```
-
-```{r, wos query for title only searches; genus}
-create_query_string_T_wos_genus <- function(genus, synonyms, additionalkeywords){
-  if (missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TI = "', genus, '"'))
-  } 
-  if (!missing(additionalkeywords) & missing(synonyms)) {
-    return(paste0('TI = ("', genus, '" AND ', additionalkeywords, ')'))
-  }
-  if (missing(additionalkeywords) & !missing(synonyms)) {
-    temp_string <- paste0('TI = ("', genus, '" OR "', synonyms[1], '")')
-    if (length(synonyms)==1) {
-      return(paste0(temp_string))
[…]
