+++ man/YearsPublishing.Rd
@@ -19,7 +19,6 @@ The number of years since the first publication in relation to the species.
 \dontrun{
 Woylie <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Bettongia", species = "penicillata")
 }
 YearsPublishing(Woylie)
