+++ man/Allindices.Rd
@@ -25,7 +25,6 @@ This function returns a dataframe of the summary of all of the indices.
 \dontrun{
 Woylie <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Bettongia", species = "penicillata")
 }
 Allindices(Woylie,
