+++ man/SpHindex.Rd
@@ -19,7 +19,6 @@ This function calculates the h-index of a species.
 \dontrun{
 Woylie <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Bettongia", species = "penicillata")
 }
 SpHindex(Woylie)
