+++ man/SpH5.Rd
@@ -19,7 +19,6 @@ This function calculates the h-index of a species in the past 5 years.
 \dontrun{
 Woylie <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Bettongia", species = "penicillata")
 }
 SpH5(Woylie)
