+++ man/SpHAfterdate.Rd
@@ -21,7 +21,6 @@ This function calculates the h-index using a given date up till the newest recor
 \dontrun{
 Woylie <- Fetch(db = "scopus",
                 search = "tak",
-                level = "species",
                 genus = "Bettongia", species = "penicillata")
 }
 SpHAfterdate(Woylie, "2000-01-01")
