### 2022-03-02 197eb0e
added some error messages to Count() and Fetch()
