+++ specieshindex.Rmd
@@ -541,6 +541,15 @@ sppub_plot_theme <- function() {
 ```{r}
 #count() wrapper
 Count <- function(db, search, genus, species = NULL, synonyms, additionalkeywords) {
+  if (missing(db)) {
+    stop('Pick a database by setting db = "scopus" / "wos" / "base".')
+  }
+  if (missing(search)) {
+    stop('Pick the search field(s) by setting search = "t" for title only or search = "tak" for title, absract, and keywords')
+  }
+  if (missing(genus)) {
+    stop('Genus is missing from your query.')
+  }
   if (db == "scopus" & search == "t") {
     countsp <- Count_scopus(search = "t", genus, species, synonyms, additionalkeywords)
   } else if (db == "scopus" & search == "tak") {
@@ -561,6 +570,15 @@ Count <- function(db, search, genus, species = NULL, synonyms, additionalkeyword
 ```{r}
 #fetch() wrapper
 Fetch <- function(db, search, genus, species = NULL, synonyms, additionalkeywords, language = 0) {
+  if (missing(db)) {
+    stop('Pick a database by setting db = "scopus" / "wos" / "base".')
+  }
+  if (missing(search)) {
+    stop('Pick the search field(s) by setting search = "t" for title only or search = "tak" for title, absract, and keywords')
+  }
+  if (missing(genus)) {
+    stop('Genus is missing from your query.')
+  }
   if (db == "scopus" & search == "t") {
     fetchsp <- FetchT_scopus(genus, species, synonyms, additionalkeywords, language)
   } else if (db == "scopus" & search == "tak") {
