+++ README.rmd
@@ -2,7 +2,7 @@
 output: github_document
 ---
 
-<img src="README_files/figure-gfm/stickerfile.png" alt="hexsticker" height="250px" align="right" />
+<img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" />
 
 # specieshindex
 
