### 2022-03-02 a048d31
turning back on a couple tests
