### 2022-03-02 2ea98d6
more string construction tests for #51
