+++ tests/testthat/test_base.R
@@ -39,6 +39,10 @@ test_that("string construction", {
   expect_type(specieshindex:::create_query_string_T_wos("Bettongia",additionalkeywords="cons",synonyms="Woylie"),"character")
   expect_type(specieshindex:::create_query_string_T_scopus("Bettongia"),"character")
   expect_type(specieshindex:::create_query_string_T_scopus("Bettongia",additionalkeywords="cons",synonyms="Woylie"),"character")
+  expect_type(specieshindex:::create_query_string_TAK_base("Bettongia"),"character")
+  expect_type(specieshindex:::create_query_string_TAK_base("Bettongia",additionalkeywords="cons",synonyms="Woylie"),"character")
+  expect_type(specieshindex:::create_query_string_TAK_wos("Bettongia"),"character")
+  expect_type(specieshindex:::create_query_string_TAK_wos("Bettongia",additionalkeywords="cons",synonyms="Woylie"),"character")
   
 })
   
