+++ README.rmd
@@ -2,7 +2,7 @@
 output: github_document
 ---
 
-<img src="man/figures/stickerfile.png" alt="hexsticker" height="120px" align="right" />
+<img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" />
 
 # specieshindex
 
@@ -162,9 +162,11 @@ CombineSp
 
 Once you are happy with your dataset, you can make some nice plots. Using `plotAllindices()`, we can compare the indices against each other.
 
-```{r, fig.align='center'}
+```{r, eval=FALSE}
 plotAllindices(CombineSp)
 ```
+
+<img src="man/figures/unnamed-chunk-11-1.png" alt="h100" align="centre" />
 
 **Figure 1.** The *h*-index, *m*-index, *i10* index, and *h5* index of the Woylie, Platypus, Koala, and Quokka.
 
@@ -172,7 +174,7 @@ plotAllindices(CombineSp)
 
 You can also visualise the total publication per year with `getYear()` and `plotPub()`.
 
-```{r, fig.align='center'}
+```{r, eval=FALSE}
 extract_year_W <- getYear(data = Woylie,
                           genus = "Bettongia", species = "penicillata")
 extract_year_Q <- getYear(data = Quokka,
@@ -184,6 +186,8 @@ extract_year_K <- getYear(data = Koala,
 Combine_pub <- rbind(extract_year_W, extract_year_Q, extract_year_P, extract_year_K)
 plotPub(Combine_pub)
 ```
+
+<img src="man/figures/unnamed-chunk-12-1.png" alt="h100" align="centre" />
 
 **Figure 2.** The total number of publications per year of the Woylie, Platypus, Koala, and Quokka.
 
