+++ R/specieshindex.R
@@ -4,10 +4,10 @@
 #' @title Search count of literature
 #'
 #' @param db Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").
-#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
+#' @param search Search fields. Title only ("t") or title, abstract, or keywords ("tak").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
-#' @param synonyms Alternate species names.
+#' @param synonyms Alternate species names; optional.
 #' @param additionalkeywords Optional search terms.
 #'
 #' @return Search count of the genus or species with the given \code{genus} and/or \code{species}.
@@ -66,10 +66,10 @@ Count <- function(db,
 #' @title Fetch citation records
 #'
 #' @param db Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").
-#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
+#' @param search Search fields. Title only ("t") or title, abstract, or keywords ("tak").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
-#' @param synonyms Alternate species names.
+#' @param synonyms Alternate species names; optional.
 #' @param additionalkeywords Optional search terms.
 #' @param language Language of the paper; default is 0, enter 1 to retrieve the variable. Scopus only.
 #'
@@ -127,7 +127,7 @@ Fetch <- function(db,
 #' 
 #' @title Search count from Scopus 
 #'
-#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
+#' @param search Search fields. Title only ("t") or title, abstract, or keywords ("tak").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
 #' @param synonyms Alternate species names.
@@ -182,7 +182,7 @@ Count_scopus <- function(search,
                                                                               additionalkeywords),
                                        httpAccept = "application/xml")) #format the URL to be sent to the API
   } else {
-    stop('Set search = "t" for title-only searches, or "tak" for searches in the title, abstract, and keywords.')
+    stop('Set search = "t" for title-only searches, or "tak" for searches in the title, abstract, or keywords.')
   }
   httr::stop_for_status(response) #pass any HTTP errors to the R console
   response_data <- XML::xmlParse(response) #parse the data to extract values
@@ -519,7 +519,7 @@ FetchTAK_scopus <- function(genus,
 #' 
 #' @title Search count from Web of Science
 #'
-#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
+#' @param search Search fields. Title only ("t") or title, abstract, or keywords ("tak").
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
 #' @param synonyms Alternate species names.
@@ -565,7 +565,7 @@ Count_wos <- function(search,
                                                                  additionalkeywords),
                              sid = sid) 
   } else {
-    stop('Set search = "t" for title-only searches, or "tak" for searches in the title, abstract, and keywords.')
+    stop('Set search = "t" for title-only searches, or "tak" for searches in the title, abstract, or keywords.')
   }
   return(count)
 }
@@ -707,7 +707,7 @@ FetchTAK_wos <- function(genus,
 #' 
 #' @title Search count from BASE
 #'
-#' @param search Search fields. Title only ("t") or title, abstract, and keywords ("tak").
[…]
