+++ README.rmd
@@ -2,9 +2,8 @@
 output: github_document
 ---
 
-<<<<<<< HEAD
-<img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" />
-=======
+<!-- <img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" /> -->
+
 ```{r, echo = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
@@ -13,8 +12,6 @@ knitr::opts_chunk$set(
 )
 ```
 
-<img src="man/figures/stickerfile.png" alt="hexsticker" height="120px" align="right" />
->>>>>>> 2ea98d6df52d101faad6205c2e6bdbc7803c2234
 
 # specieshindex
 
@@ -120,7 +117,7 @@ Count(db = "scopus",
       search = "t",
       genus = "Bettongia", species = "penicillata")
 
-#Title, abstract, and keywords; genus level
+#Title, abstract, or keywords; genus level
 Count(db = "scopus",
       search = "tak",
       genus = "Bettongia")
@@ -178,8 +175,6 @@ Once you are happy with your dataset, you can make some nice plots. Using `plotA
 plotAllindices(CombineSp)
 ```
 
-<img src="man/figures/unnamed-chunk-11-1.png" alt="h100" align="centre" />
-
 **Figure 1.** The *h*-index, *m*-index, *i10* index, and *h5* index of the Woylie, Platypus, Koala, and Quokka.
 
 <br/>
@@ -198,8 +193,6 @@ extract_year_K <- getYear(data = Koala,
 Combine_pub <- rbind(extract_year_W, extract_year_Q, extract_year_P, extract_year_K)
 plotPub(Combine_pub)
 ```
-
-<img src="man/figures/unnamed-chunk-12-1.png" alt="h100" align="centre" />
 
 **Figure 2.** The total number of publications per year of the Woylie, Platypus, Koala, and Quokka.
 

[…]
