+++ README.rmd
@@ -2,7 +2,7 @@
 output: github_document
 ---
 
-<!-- <img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" /> -->
+<img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" />
 
 ```{r, echo = FALSE}
 knitr::opts_chunk$set(
@@ -174,6 +174,8 @@ Once you are happy with your dataset, you can make some nice plots. Using `plotA
 plotAllindices(CombineSp)
 ```
 
+<img src="man/figures/unnamed-chunk-11-1.png" alt="h100" align="centre" />
+
 **Figure 1.** The *h*-index, *m*-index, *i10* index, and *h5* index of the Woylie, Platypus, Koala, and Quokka.
 
 <br/>
@@ -192,6 +194,8 @@ extract_year_K <- getYear(data = Koala,
 Combine_pub <- rbind(extract_year_W, extract_year_Q, extract_year_P, extract_year_K)
 plotPub(Combine_pub)
 ```
+
+<img src="man/figures/unnamed-chunk-12-1.png" alt="h100" align="centre" />
 
 **Figure 2.** The total number of publications per year of the Woylie, Platypus, Koala, and Quokka.
 
