+++ R/specieshindex.R
@@ -17,12 +17,14 @@
 #' \dontrun{
 #' Count(db = "scopus",
 #'       search = "t",
-#'       genus = "Osphranter", species = "rufus")
+#'       genus = "Osphranter",
+#'       species = "rufus")
 #' }
 #' \dontrun{
 #' Count(db = "scopus",
 #'       search = "t",
-#'       genus = "Osphranter", species = "rufus",
+#'       genus = "Osphranter",
+#'       species = "rufus",
 #'       synonyms = "Macropus rufus",
 #'       additionalkeywords = "conserv*")
 #' }
@@ -43,17 +45,41 @@ Count <- function(db,
     stop('Genus is missing from your query.')
   }
   if (db == "scopus" & search == "t") {
-    countsp <- Count_scopus(search = "t", genus, species, synonyms, additionalkeywords)
+    countsp <- Count_scopus(search = "t",
+                            genus,
+                            species,
+                            synonyms,
+                            additionalkeywords)
   } else if (db == "scopus" & search == "tak") {
-    countsp <- Count_scopus(search = "tak", genus, species, synonyms, additionalkeywords)
+    countsp <- Count_scopus(search = "tak",
+                            genus, 
+                            species,
+                            synonyms,
+                            additionalkeywords)
   } else if (db == "wos" & search == "t") {
-    countsp <- Count_wos(search = "t", genus, species, synonyms, additionalkeywords)
+    countsp <- Count_wos(search = "t",
+                         genus,
+                         species,
+                         synonyms,
+                         additionalkeywords)
   } else if (db == "wos" & search == "tak") {
-    countsp <- Count_wos(search = "tak", genus, species, synonyms, additionalkeywords)
+    countsp <- Count_wos(search = "tak",
+                         genus,
+                         species,
+                         synonyms,
+                         additionalkeywords)
   } else if (db == "base" & search == "t") {
-    countsp <- Count_base(search = "t", genus, species, synonyms, additionalkeywords)
+    countsp <- Count_base(search = "t",
+                          genus,
+                          species,
+                          synonyms,
+                          additionalkeywords)
   } else if (db == "base" & search == "tak") {
-    countsp <- Count_base(search = "tak", genus, species, synonyms, additionalkeywords)
+    countsp <- Count_base(search = "tak",
+                          genus,
+                          species,
+                          synonyms,
+                          additionalkeywords)
   } 
   return(countsp)
 }
@@ -85,7 +111,8 @@ Count <- function(db,
 #' \dontrun{
 #' Fetch(db = "scopus",
 #'       search = "t",
-#'       genus = "Osphranter", species = "rufus",
+#'       genus = "Osphranter",
+#'       species = "rufus",
 #'       synonyms = "Macropus rufus",
 #'       additionalkeywords = "conserv*")
 #' }
@@ -107,13 +134,26 @@ Fetch <- function(db,
     stop('Genus is missing from your query.')
   }
   if (db == "scopus" & search == "t") {
-    fetchsp <- FetchT_scopus(genus, species, synonyms, additionalkeywords, language)
+    fetchsp <- FetchT_scopus(genus,
+                             species,
+                             synonyms,
+                             additionalkeywords,
+                             language)
[…]
