### 2022-03-03 8f62732
fixing size of sticker on the website
