+++ README.rmd
@@ -4,7 +4,7 @@ output: github_document
 
 <img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" />
 
-```{r, echo = FALSE, fig.height=5}
+```{r, echo = FALSE, fig.height=3}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
