### 2022-03-03 7e2c1c3
added more tests for string construction
