+++ tests/testthat/test_wos.R
@@ -1,3 +1,31 @@
+#string construction
+test_that("string construction", {  
+  expect_type(specieshindex:::create_query_string_T_wos("Bettongia"),
+              "character")
+  expect_type(specieshindex:::create_query_string_T_wos("Bettongia",
+                                                        synonyms = "Woylie"),
+              "character")
+  expect_type(specieshindex:::create_query_string_T_wos("Bettongia",
+                                                        additionalkeywords = "cons"),
+              "character")
+  expect_type(specieshindex:::create_query_string_T_wos("Bettongia",
+                                                        additionalkeywords = "cons",
+                                                        synonyms = "Woylie"),
+              "character")
+  expect_type(specieshindex:::create_query_string_TAK_wos("Bettongia"),
+              "character")
+  expect_type(specieshindex:::create_query_string_TAK_wos("Bettongia",
+                                                           synonyms = "Woylie"),
+              "character")
+  expect_type(specieshindex:::create_query_string_TAK_wos("Bettongia",
+                                                           additionalkeywords = "cons"),
+              "character")
+  expect_type(specieshindex:::create_query_string_TAK_wos("Bettongia",
+                                                          additionalkeywords = "cons",
+                                                          synonyms = "Woylie"),
+              "character")
+})
+
 #TITLE ONLY
 # sid <- wosr::auth(username = NULL, password = NULL)
 # t_wos <- httr::GET("http://search.webofknowledge.com/esti/wokmws/ws/WokSearch-47b6e0-POST")
