### 2022-03-03 539f5f0
Fix logo size in pkgdown website
