+++ README.rmd
@@ -2,7 +2,7 @@
 output: github_document
 ---
 
-<img src="man/figures/stickerfile.png" alt="hexsticker" height="250px" align="right" />
+<img src="man/figures/stickerfile.png" alt="hexsticker" width="120" align="right" />
 
 ```{r, echo = FALSE, fig.height=3}
 knitr::opts_chunk$set(
