### 2022-03-04 b580a78
adding tests for Count functions of different db
