+++ tests/testthat/test_wrapper.R
@@ -1,5 +1,5 @@
 #count()
-test_that("count wrapper", {  
+test_that("count wrapper errors", {  
   expect_error(Count(search = "t",
                      genus = "Bettongia"))
   expect_error(Count(db = "scopus",
@@ -9,7 +9,7 @@ test_that("count wrapper", {
 })
 
 #fetch()
-test_that("fetch wrapper", {  
+test_that("fetch wrapper errors", {  
   expect_error(Fetch(search = "t",
                      genus = "Bettongia"))
   expect_error(Fetch(db = "scopus",
@@ -17,3 +17,4 @@ test_that("fetch wrapper", {
   expect_error(Fetch(db = "scopus",
                      search = "t"))
 })
+
