### 2022-03-06 6347dde
adding tests for errors
