+++ man/Count.Rd
@@ -9,13 +9,13 @@ Count(db, search, genus, species = NULL, synonyms, additionalkeywords)
 \arguments{
 \item{db}{Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").}
 
-\item{search}{Search fields. Title only ("t") or title, abstract, and keywords ("tak").}
+\item{search}{Search fields. Title only ("t") or title, abstract, or keywords ("tak").}
 
 \item{genus}{Genus classification from the binomial name.}
 
 \item{species}{Species classification from the binomial name.}
 
-\item{synonyms}{Alternate species names.}
+\item{synonyms}{Alternate species names; optional.}
 
 \item{additionalkeywords}{Optional search terms.}
 }
@@ -30,12 +30,14 @@ A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the 
 \dontrun{
 Count(db = "scopus",
       search = "t",
-      genus = "Osphranter", species = "rufus")
+      genus = "Osphranter",
+      species = "rufus")
 }
 \dontrun{
 Count(db = "scopus",
       search = "t",
-      genus = "Osphranter", species = "rufus",
+      genus = "Osphranter",
+      species = "rufus",
       synonyms = "Macropus rufus",
       additionalkeywords = "conserv*")
 }
