### 2022-03-06 21852e7
adding tests for Fetch functions
