+++ tests/testthat/test_wos.R
@@ -47,6 +47,15 @@ with_mock_api({
   })
 })
 
+#Fetch()
+with_mock_api({
+  test_that("wos scopus works", {
+    expect_POST(FetchT_wos(genus = "Bettongia"),
+               'http://search.webofknowledge.com/esti/wokmws/ws/WokSearch <soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/"')
+    
+  })
+})
+
 #TITLE ONLY
 # sid <- wosr::auth(username = NULL, password = NULL)
 # t_wos <- httr::GET("http://search.webofknowledge.com/esti/wokmws/ws/WokSearch-47b6e0-POST")
