### 2022-03-07 c2398ef
adding more tests to fetch functions
