+++ R/plots_again.R
@@ -488,7 +488,7 @@ map_lc <- ggplot(data = world) + #LC
              size = 3,
              alpha = 0.5) + 
   coord_sf(expand = FALSE) +
-  labs(title = "(a) Distribution of non-threatened species",
+  labs(title = "(a) Non-threatened species",
        x = "Longitude",
        y = "Latitude",
        colour = expression(bold(paste("species ", italic(h), "-index")))) +
@@ -508,7 +508,7 @@ map_endangered <- ggplot(data = world) + #endangered
              size = 3,
              alpha = 0.5) + 
   coord_sf(expand = FALSE) +
-  labs(title = "(b) Distribution of threatened species",
+  labs(title = "(b) Threatened species",
        x = "Longitude",
        y = "Latitude",
        colour = expression(bold(paste("species ", italic(h), "-index")))) +
