+++ .github/workflows/test-coverage.yml
@@ -13,6 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      Elsevier_API: ${{ secrets.api_test }}
 
     steps:
       - uses: actions/checkout@v2
