### 2022-04-06 22e9e93
moving retrieve api key block before tests
