+++ .github/workflows/test-coverage.yml
@@ -21,15 +21,17 @@ jobs:
         with:
           use-public-rspm: true
 
-      - uses: r-lib/actions/setup-r-dependencies@v1
-        with:
-          extra-packages: covr
-
-      - name: Test coverage
-        run: covr::codecov()
-        shell: Rscript {0}
-        
       - name: Retrieve API key from .Renviron file
         run: |
           echo MY_SECRET="$api_test" >> ~/.Renviron
           shell: bash
+      
+      - uses: r-lib/actions/setup-r-dependencies@v1
+        with:
+          extra-packages: covr
+      
+      - name: Test coverage
+        run: covr::codecov()
+        shell: Rscript {0}
+        
+      
