+++ .github/workflows/test-coverage.yml
@@ -13,6 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      Elsevier_API: ${{ secrets.api_test }}
 
     steps:
       - uses: actions/checkout@v2
@@ -23,9 +24,8 @@ jobs:
 
       - name: Retrieve API key from .Renviron file
         run: |
-          echo MY_SECRET="$api_test" >> ~/.Renviron
-        env: 
-          Elsevier_API: ${{ secrets.api_test }}
+          echo Elsevier_API ="$api_test" >> ~/.Renviron
+          shell: bash
       
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
