+++ .github/workflows/test-coverage.yml
@@ -22,14 +22,14 @@ jobs:
         with:
           use-public-rspm: true
 
+      - uses: r-lib/actions/setup-r-dependencies@v1
+        with:
+          extra-packages: covr
+          
       - name: Retrieve API key from .Renviron file
         run: |
           echo Elsevier_API ="$api_test" >> ~/.Renviron
         shell: bash
-      
-      - uses: r-lib/actions/setup-r-dependencies@v1
-        with:
-          extra-packages: covr
       
       - name: Test coverage
         run: covr::codecov()
