+++ tests/testthat/test_scopus.R
@@ -70,9 +70,9 @@ with_mock_api({
 
 with_mock_api({
   test_that("scopus query working", {
-    expect_GET(specieshindex:::scopus_request_t(request = '" AND PUBYEAR > 2019"'),
+    expect_GET(specieshindex:::scopus_request_t(request = '" AND PUBYEAR > 2019"', api_key = Elsevier_API),
                "https://api.elsevier.com/content/search/scopus?query=TITLE%28%28%22%20%22%20OR%20%22%22%29%20AND%2")
-    expect_GET(specieshindex:::scopus_request_tak(request = '" AND PUBYEAR > 2019"'),
+    expect_GET(specieshindex:::scopus_request_tak(request = '" AND PUBYEAR > 2019"', api_key = Elsevier_API),
                "https://api.elsevier.com/content/search/scopus?query=TITLE-ABS-KEY%28%28%22%20%22%20OR%20%22%22%29%20AND%20%29%22%20AND%20PUBYEAR%20%3E%202019%22&APIKey=442b9048417ef20cf680a0ae26ee4d86&count=200&start=0&view=STANDARD")
   })
 })
