+++ .github/workflows/test-coverage.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      Elsevier_API: ${{ secrets.api_test }}
+      Elsevier_API: ${{ secrets.ELSEVIER_API }}
 
     steps:
       - uses: actions/checkout@v2
