+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.2
+pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
   specieshindex: specieshindex.html
-last_built: 2022-04-22T04:40Z
+last_built: 2022-04-25T23:11Z
 
