### 2022-04-26 e1a1c1a
adding tests for fetch wrapper and Allindices()
