### 2022-04-26 1850fc3
adding tests for fetch scopus
