### 2022-04-26 916967d
adding tests for Allindices()
