### 2022-05-10 #3776 @jessicatytam
We have completed all of the tasks and addressed the comments, with the exception of adding tests, it's currently ~50% so we will probably need until the end of May if that's ok.

What would be the ideal % for test coverage?
