+++ tests/testthat/_snaps/woylie.md
@@ -1,0 +1,10 @@
+# Allindices
+
+    Code
+      Allindices(Woylie, "Bettongia", "penicillata")
+    Output
+                genus_species     species     genus publications citations journals
+      1 Bettongia penicillata penicillata Bettongia          113      1903       55
+        years_publishing  h     m i10 h5
+      1               45 26 0.578  54  4
+
