+++ R/specieshindex.R
@@ -1173,7 +1173,6 @@ SpHAfterdate <- function(data, date) {
 #' @param data The dataframe generated from \code{\link{Fetch}}.
 #' @param genus Genus classification from the binomial name.
 #' @param species Species classification from the binomial name.
-#' @param sourcetype Source type; default is 0, enter 1 to add SourceType variables.
 #'
 #' @return A datarame of all of the indices in the package.
 #' @export
@@ -1187,41 +1186,12 @@ SpHAfterdate <- function(data, date) {
 #' }
 #' Allindices(Woylie,
 #'            genus = "Bettongia",
-#'            species = "penicillata",
-#'            sourcetype = 0)
+#'            species = "penicillata",)
 #' 
 Allindices <- function(data,
                        genus,
-                       species,
-                       sourcetype = 0) {
-  if (sourcetype == 1 & all.equal(0, data$citations) == FALSE) {
-    combine <- data.frame(paste0(genus, " ", species),
-                          paste0(species),
-                          paste0(genus),
-                          TotalPub(data),
-                          TotalCite(data),
-                          TotalJournals(data),
-                          YearsPublishing(data),
-                          SpHindex(data),
-                          SpMindex(data),
-                          Spi10(data),
-                          SpH5(data))
-    combine[is.na(combine)] <- 0 #replace NA values with 0
-    combine_st <- cbind(combine, SourceType(data))
-    colnames(combine_st) <- c("genus_species",
-                              "species",
-                              "genus",
-                              "publications",
-                              "citations",
-                              "journals",
-                              "years_publishing",
-                              "h",
-                              "m",
-                              "i10",
-                              "h5",
-                              names(SourceType(data)))
-    return(combine_st)
-  } else if (all.equal(0, data$citations) == TRUE) {
+                       species) {
+  if (all.equal(0, data$citations) == TRUE) {
     zeroIndex <- data.frame(genus_species = paste0(genus, " ", species),
                             species = paste0(species),
                             genus = paste0(genus),
