### 2022-05-26 #3776 @jessicatytam
We've got through adding lots of tests and almost everything is now tested except for one big issue.

We're dealing with some issues with testing the functions that query an API multiple times. Since there are limitations in the scopus API, we need >10 GET requests in a single function to retrieve all of the data while handling edge cases smoothly. We are using `httptest::with_mock_api()` to test the API functions currently, but `with_mock_api()` can only test 1 GET request per test. So to test everything in those functions is difficult given the combination of the constraints of the Scopus API and the `mock_api` framework. So we're a bit stuck on that unless you can see a clever solution?

Here's the link the codecov: https://app.codecov.io/gh/jessicatytam/specieshindex/blob/master/R/specieshindex.R .  Check out the function `FetchT_scopus` which is being tested but the test exits (successfully) after the first GET call.  Any help much appreciated.
