### 2022-05-26 8bc6f60
more query string tests
