### 2022-05-26 1f95b03
updating example to remove argument for allindices
