### 2022-05-30 eb318d2
Reduce code duplication by using switch
