### 2022-05-30 3e86a88
Reduce number of if/else by passing search instead of a fixed string
