### 2022-05-30 #3776 @Bisaloo
@jessicatytam, while investigating your problem, I had another look and I want to already now mention that @DrMattG's point about code duplication is not yet properly addressed in my opinion.

Most functions still include a lot of duplicated code. I have submitted a pull request (https://github.com/jessicatytam/specieshindex/pull/60) where I show how you could reduce code duplication. Please have a look and try to apply this kind of fixes throughout the code. Let me know if it's not clear.

On the topic of duplication, I notice that the documentation of example datasets is duplicated as well. You documented them in each individual files (`Koala.R`, `Platypus.R`, etc.) and at the end of the main `specieshindex.R` file. Could you have a look please?

I would like to emphasize that this insistence on avoiding code duplication is not just to annoy you. Code with high duplication is very hard / impossible to maintain in the long run. If you refactor the code to reduce duplication, future updates will require change in a single place, instead of throughout the codebase.
