### 2022-05-31 jessicatytam/specieshindex#62 @Bisaloo
Do not grow data.frame inside a loop
Relevant blog post: https://privefl.github.io/blog/why-loops-are-slow-in-r/

Affected functions: `FetchT_scopus()` & `FetchTAK_scopus()`

Your case is slightly more complicated than the usual basic example because you don't know the size of the result in advance. As mentioned in the blog post, in this case, you should use a list to store your results (this works because of differences in the way R stores list vs data.frames internally). You can then convert this list into a data.frame **after** the loop has finished running with `do.call(rbind, list)` or `dplyr::bind_rows(list)`.
