### 2022-06-15 #3776 @jessicatytam
- [ ] unused arguments (https://github.com/jessicatytam/specieshindex/issues/61#issue-1253599082)
- [ ] storing results w/in a loop (https://github.com/jessicatytam/specieshindex/issues/62#issue-1253657243)
- [ ] reduce code duplication (https://github.com/jessicatytam/specieshindex/pull/60#issue-1252587022)
- [ ] fix dataset documentation duplication
