### 2022-06-16 70fa605
removed duplicated documentation and replaced print() with message()
