+++ R/languages.R
@@ -1,12 +1,0 @@
-#' List of languages of documents found on Scopus.
-#' Data was retrieved on 9 March 2021.
-#' 
-#' @title Languages
-#'
-#' @format A csv file with the complete list of languages of documents found on Scopus.
-#'
-#' @source \url{https://www.elsevier.com/solutions/scopus/how-scopus-works/content}
-#' 
-#' @noRd
-#' 
-"languages"
