+++ R/Platypus.R
@@ -1,12 +1,0 @@
-#' Citation records of platypus (Ornithorhynchus anatinus) from Scopus.
-#' Data was retrieved on 10 July 2020.
-#' 
-#' @title Platypus dataset
-#'
-#' @format A data frame with 321 rows and 20 variables
-#'
-#' @source \url{http://api.elsevier.com/content/search/scopus}
-#' 
-#' @noRd
-#' 
-"Platypus"
