+++ R/specieshindex.R
@@ -276,7 +276,7 @@ FetchT_scopus <- function(genus,
                         species = paste0(species),
                         synonyms,
                         additionalkeywords) #check the number of records
-  print(paste(count, "records found."))
+  message(count, "records found.")
   if (count < 1) {
     noCitations <- data.frame(citations = 0)
     return(noCitations)
@@ -340,8 +340,8 @@ FetchT_scopus <- function(genus,
       looprepeat <- ceiling(count/step_size)-1 #the number of loop times, rounded up to the nearest integer
       #loop starts
       for (i in 0:looprepeat) { 
-        print(paste("starting iteration: ", i, " Note: iteration size is ", step_size, " records, which runs of 200 records inside each iteration."))
-        print(paste("Fetching records now."))
+        message("starting iteration: ", i, " Note: iteration size is ", step_size, " records, which runs of 200 records inside each iteration.")
+        message("Fetching records now.")
         search <- rscopus::scopus_search(query = create_query_string_T_scopus(genus,
                                                                               species = paste0(species),
                                                                               synonyms,
@@ -388,7 +388,7 @@ FetchT_scopus <- function(genus,
       }
   }
   returned <- dim(datalist)[1]
-  print(paste(returned, "records retrived in total."))
+  message(returned, "records retrived in total.")
   return(datalist)
 }
 
@@ -444,7 +444,7 @@ FetchTAK_scopus <- function(genus,
                         species = paste0(species),
                         synonyms,
                         additionalkeywords) #check the number of records
-  print(paste(count, "records found."))
+  message(count, "records found.")
   if (count < 1) {
     noCitations <- data.frame(citations = 0)
     return(noCitations)
@@ -508,8 +508,8 @@ FetchTAK_scopus <- function(genus,
       looprepeat <- ceiling(count/step_size)-1 #the number of loop times, rounded up to the nearest integer
       #loop starts
       for (i in 0:looprepeat) { 
-        print(paste("starting iteration: ", i, " Note: iteration size is ", step_size, " records, which runs of 200 records inside each iteration."))
-        print(paste("Fetching records now."))
+        message("starting iteration: ", i, " Note: iteration size is ", step_size, " records, which runs of 200 records inside each iteration.")
+        message("Fetching records now.")
         search <- rscopus::scopus_search(query = create_query_string_TAK_scopus(genus,
                                                                                 species = paste0(species),
                                                                                 synonyms, 
@@ -556,7 +556,7 @@ FetchTAK_scopus <- function(genus,
       }
   }
   returned <- dim(datalist)[1]
-  print(paste(returned, "records retrived in total."))
+  message(returned, "records retrived in total.")
   return(datalist)
 }
 
@@ -677,7 +677,7 @@ FetchT_wos <- function(genus,
                      species = paste0(species),
                      synonyms,
                      additionalkeywords) #check the number of records
-  print(paste(count, "records found."))
+  message(count, "records found.")
   if (count < 1) {
     noCitations <- data.frame(citations = 0)
     return(noCitations)
@@ -695,7 +695,7 @@ FetchT_wos <- function(genus,
   names(results)[names(results) == "date"] <- "cover_date"
   #showing final list of records
   returned <- nrow(results)
[…]
