+++ R/Quokka.R
@@ -1,12 +1,0 @@
-#' Citation records of quokka (Setonix brachyurus) from Scopus.
-#' Data was retrieved on 10 July 2020.
-#' 
-#' @title Quokka dataset
-#'
-#' @format A data frame with 242 rows and 20 variables
-#'
-#' @source \url{http://api.elsevier.com/content/search/scopus}
-#' 
-#' @noRd
-#' 
-"Quokka"
