+++ R/Woylie.R
@@ -1,12 +1,0 @@
-#' Citation records of woylie (Bettongia penicillata) from Scopus.
-#' Data was retrieved on 10 July 2020.
-#' 
-#' @title Woylie dataset
-#'
-#' @format A data frame with 113 rows and 20 variables
-#'
-#' @source \url{http://api.elsevier.com/content/search/scopus}
-#' 
-#' @noRd
-#' 
-"Woylie"
