+++ tests/testthat/test_woylie.R
@@ -10,8 +10,6 @@ test_that("Woylie_stats", {
   expect_true(Spi10(Woylie)>=54)
   expect_true(SpH5(Woylie)>=4)
   expect_true(SpHAfterdate(Woylie, "2000-01-01")>=20)
-  expect_type(specieshindex:::sp_check("Bettongia", "penicillata"),
-              "character")
 })
 
 test_that("Allindices", {
