+++ R/Koala.R
@@ -1,12 +1,0 @@
-#' Citation records of koala (Phascolarctos cinereus) from Scopus.
-#' Data was retrieved on 10 July 2020.
-#' 
-#' @title Koala dataset
-#'
-#' @format A data frame with 773 rows and 20 variables
-#'
-#' @source \url{http://api.elsevier.com/content/search/scopus}
-#' 
-#' @noRd
-#' 
-"Koala"
