### 2022-08-23 #3776 @Bisaloo
@openjournals/joss-eics, can we withdraw this submission please?
